Crash

Documentation for StochasticPrograms.jl's crash methods, for use in structured solvers and sample-based solvers.

Index

StochasticPrograms.Crash

StochasticPrograms.Crash.Custom

StochasticPrograms.Crash.EVP

StochasticPrograms.Crash.FeasiblePoint

StochasticPrograms.Crash.None

StochasticPrograms.Crash.PreSolve

StochasticPrograms.Crash.Scenario

Crash methods

StochasticPrograms.Crash — Module

Crash

Collection of crash methods used to generate initial decisions in structured algorithms.

...

Available crash methods

None

EVP

Scenario

Custom

...

Examples

The following solves a stochastic program sp created in StochasticPrograms.jl using an L-shaped algorithm with trust-region and Clp as an lpsolver and by generating an initial decision with the EVP crash.

julia> optimize!(sp, solver = LShapedSolver(GLPKSolverLP(), crash=Crash.EVP(), regularize = TrustRegion()))
L-Shaped Gap  Time: 0:00:00 (8 iterations)
  Objective:       -855.8333333333339
  Gap:             0.0
  Number of cuts:  4
  Iterations:      8
:Optimal

StochasticPrograms.Crash.Custom — Type

Custom(x₀)

Use the user-supplied x₀ as initial decision.

StochasticPrograms.Crash.EVP — Type

EVP

Solve the expected value problem corresponding to the stochastic program and use the expected value solution as initial decision.

StochasticPrograms.Crash.FeasiblePoint — Type

FeasiblePoint

Generate a feasible first-stage decision as initial decision.

StochasticPrograms.Crash.None — Type

None

Randomize the initial decision (default).

StochasticPrograms.Crash.PreSolve — Type

PreSolve

Run a provided (ideally suboptimal) optimization procedure and use the (sub)optimal solution as initial decision.

StochasticPrograms.Crash.Scenario — Type

Scenario

Solve the wait-and-see problem corresponding a supplied scenario and use the optimal solution as initial decision.