L-shaped solvers

Documentation for StochasticPrograms.jl's L-shaped solvers.

Index

StochasticPrograms.LShapedSolvers.Aggregate

StochasticPrograms.LShapedSolvers.AsynchronousExecution

StochasticPrograms.LShapedSolvers.ClusterAggregate

StochasticPrograms.LShapedSolvers.ClusterAggregation

StochasticPrograms.LShapedSolvers.ClusterByReference

StochasticPrograms.LShapedSolvers.Consolidate

StochasticPrograms.LShapedSolvers.Consolidation

StochasticPrograms.LShapedSolvers.DontAggregate

StochasticPrograms.LShapedSolvers.DontConsolidate

StochasticPrograms.LShapedSolvers.DontRegularize

StochasticPrograms.LShapedSolvers.DynamicAggregate

StochasticPrograms.LShapedSolvers.DynamicAggregation

StochasticPrograms.LShapedSolvers.Hierarchical

StochasticPrograms.LShapedSolvers.HybridAggregate

StochasticPrograms.LShapedSolvers.HybridAggregation

StochasticPrograms.LShapedSolvers.Kmedoids

StochasticPrograms.LShapedSolvers.LShaped

StochasticPrograms.LShapedSolvers.LShapedSolver

StochasticPrograms.LShapedSolvers.LV

StochasticPrograms.LShapedSolvers.LevelSet

StochasticPrograms.LShapedSolvers.NoAggregation

StochasticPrograms.LShapedSolvers.NoConsolidation

StochasticPrograms.LShapedSolvers.NoRegularization

StochasticPrograms.LShapedSolvers.PartialAggregate

StochasticPrograms.LShapedSolvers.PartialAggregation

StochasticPrograms.LShapedSolvers.RD

StochasticPrograms.LShapedSolvers.RegularizedDecomposition

StochasticPrograms.LShapedSolvers.SelectClosest

StochasticPrograms.LShapedSolvers.SelectClosestToReference

StochasticPrograms.LShapedSolvers.SelectDecaying

StochasticPrograms.LShapedSolvers.SelectRandom

StochasticPrograms.LShapedSolvers.SelectUniform

StochasticPrograms.LShapedSolvers.SerialExecution

StochasticPrograms.LShapedSolvers.StaticCluster

StochasticPrograms.LShapedSolvers.SynchronousExecution

StochasticPrograms.LShapedSolvers.TR

StochasticPrograms.LShapedSolvers.TrustRegion

StochasticPrograms.LShapedSolvers.FullAggregation

StochasticPrograms.LShapedSolvers.absolute_distance

StochasticPrograms.LShapedSolvers.angular_distance

StochasticPrograms.LShapedSolvers.at_tolerance

StochasticPrograms.LShapedSolvers.spatioangular_distance

L-shaped solver factory

StochasticPrograms.LShapedSolvers.LShapedSolver — Type

LShapedSolver(lpsolver::AbstractMathProgSolver; <keyword arguments>)

Return an L-shaped algorithm object that can optimize a two-stage StochasticPrograms. Supply lpsolver, a MathProgBase solver capable of solving linear-quadratic problems.

The following L-shaped regularizations are available

NoRegularization: L-shaped algorithm (default)

RegularizedDecomposition: Regularized decomposition ?RegularizedDecomposition for parameter descriptions.

TrustRegion: Trust-region ?TrustRegion for parameter descriptions.

LevelSet: Level-set ?LevelSet for parameter descriptions.

The following aggregation schemes are available

NoAggregation: Multi-cut L-shaped algorithm (default)

PartialAggregation: ?PartialAggregation for parameter descriptions.

FullAggregation: ?FullAggregation for parameter descriptions.

DynamicAggregation: ?DynamicAggregation for parameter descriptions.

ClusterAggregation: ?ClusterAggregation for parameter descriptions.

HybridAggregation: ?HybridAggregation for parameter descriptions.

The following consolidation schemes are available

NoConsolidation

Consolidation

The following execution policies are available

Serial: Classical L-shaped (default)

Synchronous: Classical L-shaped run in parallel

Asynchronous: Asynchronous L-shaped ?Asynchronous for parameter descriptions.

...

Arguments

lpsolver::AbstractMathProgSolver: MathProgBase solver capable of solving linear (and possibly quadratic) programs.

feasibility_cuts::Bool = false: Specify if feasibility cuts should be used

subsolver::AbstractMathProgSolver = lpsolver: Optionally specify a different solver for the subproblems.

regularize::AbstractRegularizer = DontRegularize(): Specify regularization procedure (DontRegularize, RegularizedDecomposition/RD/WithRegularizedDecomposition, TrustRegion/TR/WithTrustRegion, LevelSet/LV/WithLevelSets).

aggregate::AbstractAggregator = DontAggregate(): Specify aggregation procedure (DontAggregate, Aggregate, PartialAggregate, DynamicAggregate)

consolidate::AbstractConsolidator = DontConsolidate(): Specify consolidation procedure (DontConsolidate, Consolidate)

execution::Execution = Serial: Specify how algorithm should be executed (Serial, Synchronous, Asynchronous). Distributed variants requires worker cores.

crash::CrashMethod = Crash.None: Crash method used to generate an initial decision. See ?Crash for alternatives.

<keyword arguments>: Algorithm specific parameters, See ?LShaped for list of possible arguments and default values.

...

Examples

The following solves a stochastic program sp created in StochasticPrograms.jl using the L-shaped algorithm with GLPK as an lpsolver.

julia> optimize!(sp, solver = LShapedSolver(GLPKSolverLP()))
L-Shaped Gap  Time: 0:00:00 (6 iterations)
  Objective:       -855.8333333333339
  Gap:             0.0
  Number of cuts:  8
  Iterations:      6
:Optimal

StochasticPrograms.LShapedSolvers.LShaped — Type

LShaped

Functor object for the L-shaped algorithm. Create using the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

...

Algorithm parameters

τ::AbstractFloat = 1e-6: Relative tolerance for convergence checks.

debug::Bool = false: Specifies if extra information should be saved for debugging purposes. Defaults to false for memory efficiency.

cut_scaling::AbstractFloat = 1.0: Rescaling factor for cutting planes to improve numerical stability.

log::Bool = true: Specifices if L-shaped procedure should be logged on standard output or not.

...

Execution

StochasticPrograms.LShapedSolvers.SerialExecution — Type

SerialExecution

Functor object for using serial execution in a lshaped algorithm. Create by supplying a Serial object through execution in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

StochasticPrograms.LShapedSolvers.SynchronousExecution — Type

SynchronousExecution

Functor object for using synchronous execution in an L-shaped algorithm (assuming multiple Julia cores are available). Create by supplying a Synchronous object through execution in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

StochasticPrograms.LShapedSolvers.AsynchronousExecution — Type

AsynchronousExecution

Functor object for using synchronous execution in an L-shaped algorithm (assuming multiple Julia cores are available). Create by supplying a Asynchronous object through execution in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

Regularization

StochasticPrograms.LShapedSolvers.NoRegularization — Type

NoRegularization

Empty functor object for running an L-shaped algorithm without regularization.

StochasticPrograms.LShapedSolvers.DontRegularize — Type

DontRegularize

Factory object for NoRegularization. Passed by default to regularize in the LShapedSolver factory function.

StochasticPrograms.LShapedSolvers.RegularizedDecomposition — Type

RegularizedDecomposition

Functor object for using regularized decomposition regularization in an L-shaped algorithm. Create by supplying an RD object through regularize in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

σ::AbstractFloat = 1.0: Initial value of regularization parameter. Controls the relative penalty of the deviation from the current major iterate.

σ̅::AbstractFloat = 4.0: Maximum value of the regularization parameter.

σ̲::AbstractFloat = 0.5: Minimum value of the regularization parameter.

log::Bool = true: Specifices if L-shaped procedure should be logged on standard output or not.

penaltyterm::PenaltyTerm = Quadratic: Specify penaltyterm variant (Quadratic, Linearized, InfNorm, [ManhattanNorm][@ref])

...

StochasticPrograms.LShapedSolvers.RD — Type

RD

Factory object for RegularizedDecomposition. Pass to regularize in the LShapedSolver factory function. Equivalent factory calls: RD, WithRD, RegularizedDecomposition, WithRegularizedDecomposition. See ?RegularizedDecomposition for parameter descriptions.

StochasticPrograms.LShapedSolvers.TrustRegion — Type

TrustRegion

Functor object for using trust-region regularization in an L-shaped algorithm. Create by supplying a TR object through regularize in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

γ::T = 1e-4: Relative tolerance for deciding if a minor iterate should be accepted as a new major iterate.

Δ::AbstractFloat = 1.0: Initial size of ∞-norm trust-region.

Δ̅::AbstractFloat = 1000.0: Maximum size of ∞-norm trust-region.

...

StochasticPrograms.LShapedSolvers.TR — Type

TR

Factory object for TrustRegion. Pass to regularize in the LShapedSolver factory function. Equivalent factory calls: TR, WithTR, TrustRegion, WithTrustRegion. See ?TrustRegion for parameter descriptions.

StochasticPrograms.LShapedSolvers.LevelSet — Type

LevelSet

Functor object for using level-set regularization in an L-shaped algorithm. Create by supplying an LV object through regularize in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

λ::AbstractFloat = 0.5: Controls the level position L = (1-λ)θ + λQ̃, a convex combination of the current lower and upper bound.

penaltyterm::PenaltyTerm = Quadratic: Specify penaltyterm variant (Quadratic, Linearized, InfNorm, [ManhattanNorm][@ref])

...

StochasticPrograms.LShapedSolvers.LV — Type

LV

Factory object for LevelSet. Pass to regularize in the LShapedSolver factory function. Equivalent factory calls: LV, WithLV, LevelSet, WithLevelSets. See ?LevelSet for parameter descriptions.

Aggregation

StochasticPrograms.LShapedSolvers.DontAggregate — Type

DontAggregate

Factory object for NoAggregation. Passed by default to aggregate in the LShapedSolver factory function.

StochasticPrograms.LShapedSolvers.NoAggregation — Type

NoAggregation

Empty functor object for running an L-shaped algorithm without aggregation (multi-cut L-shaped).

StochasticPrograms.LShapedSolvers.Aggregate — Type

Aggregate

Factory object for FullAggregation. Pass to aggregate in the LShapedSolver factory function.

StochasticPrograms.LShapedSolvers.PartialAggregate — Type

PartialAggregate

Factory object for PartialAggregation. Pass to aggregate in the LShapedSolver factory function. See ?PartialAggregation for parameter descriptions.

StochasticPrograms.LShapedSolvers.PartialAggregation — Type

PartialAggregation

Functor object for using partial aggregation in an L-shaped algorithm. Create by supplying a PartialAggregate object through aggregate in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

size::Int: Number of cuts in each aggregate

...

StochasticPrograms.LShapedSolvers.FullAggregation — Method

FullAggregation

Functor object for using complete aggregation in an L-shaped algorithm. Create by supplying an Aggregate object through aggregate in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

StochasticPrograms.LShapedSolvers.DynamicAggregate — Type

DynamicAggregate(naggregates::Integer, rule::SelectionRule; lock_after::Function = (τ,n)->false)

Factory object for DynamicAggregation. Pass to aggregate in the LShapedSolver factory function. See ?DynamicAggregation for parameter descriptions.

StochasticPrograms.LShapedSolvers.DynamicAggregation — Type

DynamicAggregation

Functor object for using dynamic aggregation in an L-shaped algorithm. Create by supplying a DynamicAggregate object through aggregate in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

The following selection rules are available

SelectUniform

SelectDecaying

[SelectRandom](@ref

SelectClosest

SelectClosestToReference

...

Parameters

naggregates::Int: Number of aggregates

rule::SelectionRule: Rule that determines which aggregate an incoming cut should be placed in

lock_after::Function = (τ,n)->false: Function that determines if the current aggregation scheme should be fixed, based on the current optimality gap τ and the number of iterations n

...

StochasticPrograms.LShapedSolvers.ClusterAggregate — Type

ClusterAggregate(rule::ClusterRule; lock_after::Function = (τ,n)->false)

Factory object for ClusterAggregation. Pass to aggregate in the LShapedSolver factory function. See ?ClusterAggregation for parameter descriptions.

StochasticPrograms.LShapedSolvers.ClusterAggregation — Type

ClusterAggregation

Functor object for using cluster aggregation in an L-shaped algorithm. Create by supplying a ClusterAggregate object through aggregate in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

The following cluster rules are available

StaticCluster

ClusterByReference

[Kmedoids](@ref

Hierarchical

...

Parameters

rule::ClusterRule: Rule that determines how cuts should be sorted into clusters

lock_after::Function = (τ,n)->false: Function that determines if the current aggregation scheme should be fixed, based on the current optimality gap τ and the number of iterations n

...

StochasticPrograms.LShapedSolvers.HybridAggregate — Type

HybridAggregate(initial::AbstractAggregator, final::AbstractAggregator, τ::AbstractFloat)

Factory object for HybridAggregation. Pass to aggregate in the LShapedSolver factory function. See ?HybridAggregation for parameter descriptions.

StochasticPrograms.LShapedSolvers.HybridAggregation — Type

HybridAggregation

Functor object for using hybrid aggregation in an L-shaped algorithm. Create by supplying a HybridAggregate object through aggregate in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

initial::AbstractAggregator: Initial aggregation scheme

final::AbstractAggregator: Final aggregation scheme

τ::T: The active aggregation scheme is switched from initial to final when the optimality gap decreases below τ

...

Selection rules

StochasticPrograms.LShapedSolvers.SelectClosest — Type

SelectClosest(τ::AbstractFloat; distance::Function = absolute_distance)

Incoming cuts are placed into the closest aggregate, according the supplied distance function. An empty aggregate is chosen if no aggregate is within the tolerance τ

The following distance measures are available

absolute_distance

angular_distance

[spatioangular_distance](@ref

StochasticPrograms.LShapedSolvers.SelectClosestToReference — Type

SelectClosestToReference(τ::AbstractFloat; distance::Function = absolute_distance)

Incoming cuts are placed into an aggregate based on the distance to a reference cut, according the supplied distance function. Behaves as SelectClosest if not withing the tolerance τ to the reference cut.

The following distance measures are available

absolute_distance

angular_distance

[spatioangular_distance](@ref

StochasticPrograms.LShapedSolvers.SelectDecaying — Type

SelectDecaying(T₀::Integer, T̲::Integer = 1, γ::T)

Behaves like SelectUniform, but the uniform aggregate size decays by γ each iteration, starting from T₀. T̲ is an optional lower bound on the aggregate size.

StochasticPrograms.LShapedSolvers.SelectRandom — Type

SelectRandom(max = Inf)

Incoming cuts are placed into aggregates randomly. An optional maximum number of cuts max can be specified.

StochasticPrograms.LShapedSolvers.SelectUniform — Type

SelectUniform(n::Integer)

Incoming cuts are placed into aggregates uniformly, so that each aggregate has at most n cuts. Behaves as PartialAggregation.

Cluster rules

StochasticPrograms.LShapedSolvers.ClusterByReference — Method

ClusterByReference(τ::AbstractFloat; distance::Function = absolute_distance)

Buffered cuts are aggregated if within the tolerance τ to a reference cut, according the supplied distance function. Behaves as multi-cut otherwise.

The following distance measures are available

absolute_distance

angular_distance

[spatioangular_distance](@ref

StochasticPrograms.LShapedSolvers.Hierarchical — Type

Hierarchical(nclusters::Int; distance::Function = absolute_distance, linkage::Symbol = :single)

Buffered cuts are sorted into nclusters clusters, using a Hierarchical algorithm, with the given linkage, over a generalized distance matrix.

The following distance measures are available

absolute_distance

angular_distance

[spatioangular_distance](@ref

StochasticPrograms.LShapedSolvers.Kmedoids — Type

Kmedoids(nclusters::Int; distance::Function = absolute_distance)

Buffered cuts are sorted into nclusters clusters, using a K-medoids algorithm over a generalized distance matrix.

The following distance measures are available

absolute_distance

angular_distance

[spatioangular_distance](@ref

StochasticPrograms.LShapedSolvers.StaticCluster — Type

StaticCluster(clusters::Vector{Float64})

Buffered cuts are sorting according to the supplied weights clusters

Distance measures

StochasticPrograms.LShapedSolvers.absolute_distance — Method

absolute_distance(c₁::AnyOptimalityCut, c₂::AnyOptimalityCut)

Absolute distance between two optimality cuts

StochasticPrograms.LShapedSolvers.angular_distance — Method

angular_distance(c₁::AnyOptimalityCut, c₂::AnyOptimalityCut)

Angular distance between two optimality cuts

StochasticPrograms.LShapedSolvers.spatioangular_distance — Method

spatioangular_distance(c₁::AnyOptimalityCut, c₂::AnyOptimalityCut)

Spatioangular distance between two optimality cuts.

Consolidation

StochasticPrograms.LShapedSolvers.Consolidate — Type

Consolidate

Factory object for Consolidation. Pass to consolidate in the LShapedSolver factory function. See ?Consolidation for parameter descriptions.

StochasticPrograms.LShapedSolvers.Consolidation — Type

Consolidation

Functor object for using consolidation in an L-shaped algorithm. Create by supplying a Consolidate object through consolidate in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

...

Algorithm parameters

tresh::T = 0.95: Relative amount of redundant cuts in a former iteration required to consider the iteration redundant

at::Int = 5.0: Number of times an iteration can be redundant before consolidation is triggered

rebuild::Function = at_tolerance(): Function deciding when the master model should be rebuilt according to performed consolidations

...

StochasticPrograms.LShapedSolvers.DontConsolidate — Type

DontConsolidate

Factory object for NoConsolidation. Passed by default to consolidate in the LShapedSolver factory function.

StochasticPrograms.LShapedSolvers.at_tolerance — Method

at_tolerance(τ = 0.4, miniter = 0)

Rebuild master when at least nconsolidations*miniter iterations has passed and the ratio of number of cuts in the consolidated collection and the number of cuts in the master model has decreased below τ.

StochasticPrograms.LShapedSolvers.NoConsolidation — Type

NoRegularization

Empty functor object for running the L-shaped algorithm without consolidation.