Progressive-hedging solvers

Documentation for StochasticPrograms.jl's progressive-hedging solvers.

Index

StochasticPrograms.ProgressiveHedgingSolvers.Adaptive

StochasticPrograms.ProgressiveHedgingSolvers.AdaptivePenalization

StochasticPrograms.ProgressiveHedgingSolvers.AsynchronousExecution

StochasticPrograms.ProgressiveHedgingSolvers.Fixed

StochasticPrograms.ProgressiveHedgingSolvers.FixedPenalization

StochasticPrograms.ProgressiveHedgingSolvers.ProgressiveHedging

StochasticPrograms.ProgressiveHedgingSolvers.ProgressiveHedgingSolver

StochasticPrograms.ProgressiveHedgingSolvers.SerialExecution

StochasticPrograms.ProgressiveHedgingSolvers.SynchronousExecution

Progressive-hedging solver factory

StochasticPrograms.ProgressiveHedgingSolvers.ProgressiveHedgingSolver — Type

ProgressiveHedgingSolver(qpsolver::AbstractMathProgSolver; <keyword arguments>)

Return a progressive-hedging algorithm object specified. Supply qpsolver, a MathProgBase solver capable of solving quadratic problems.

The following penalty parameter update procedures are available

Fixed: Fixed penalty (default) ?Fixed for parameter descriptions.

Adaptive: Adaptive penalty update ?Adaptive for parameter descriptions.

The following execution policies are available

Serial: Classical progressive-hedging (default)

Synchronous: Classical progressive-hedging run in parallel

Asynchronous: Asynchronous progressive-hedging ?Asynchronous for parameter descriptions.

...

Arguments

qpsolver::AbstractMathProgSolver: MathProgBase solver capable of solving quadratic programs.

penalty::AbstractPenalizer = Fixed(): Specify penalty update procedure (Fixed, Adaptive)

execution::AbstractExecuter = Serial: Specify how algorithm should be executed (Serial, Synchronous, Asynchronous). Distributed variants requires worker cores.

penaltyterm::PenaltyTerm = Quadratic: Specify penaltyterm variant (Quadratic, Linearized, InfNorm, [ManhattanNorm][@ref])

<keyword arguments>: Algorithm specific parameters, consult individual docstrings (see above list) for list of possible arguments and default values.

...

Examples

The following solves a stochastic program sp created in StochasticPrograms.jl using the progressive-hedging algorithm with Ipopt as an qpsolver.

julia> solve(sp,solver=ProgressiveHedgingSolver(IpoptSolver(print_level=0)))
Progressive Hedging Time: 0:00:06 (1315 iterations)
  Objective:  -855.8332803469432
  δ:          9.436947935542464e-7
:Optimal

StochasticPrograms.ProgressiveHedgingSolvers.ProgressiveHedging — Type

ProgressiveHedging

Functor object for the progressive-hedging algorithm. Create using the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

τ::AbstractFloat = 1e-6: Relative tolerance for convergence checks.

log::Bool = true: Specifices if progressive-hedging procedure should be logged on standard output or not.

...

Execution

StochasticPrograms.ProgressiveHedgingSolvers.SerialExecution — Type

SerialExecution

Functor object for using serial execution in a progressive-hedging algorithm. Create by supplying a Serial object through execution in the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

StochasticPrograms.ProgressiveHedgingSolvers.SynchronousExecution — Type

SynchronousExecution

Functor object for using synchronous execution in a progressive-hedging algorithm (assuming multiple Julia cores are available). Create by supplying a Synchronous object through execution in the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

StochasticPrograms.ProgressiveHedgingSolvers.AsynchronousExecution — Type

AsynchronousExecution

Functor object for using asynchronous execution in a progressive-hedging algorithm (assuming multiple Julia cores are available). Create by supplying an Asynchronous object through execution in the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

Penalties

StochasticPrograms.ProgressiveHedgingSolvers.FixedPenalization — Type

FixedPenalization

Functor object for using fixed penalty in a progressive-hedging algorithm. Create by supplying a Fixed object through penalty in the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

r::T = 1.00: Fixed penalty

...

StochasticPrograms.ProgressiveHedgingSolvers.Fixed — Type

Fixed

Factory object for FixedPenalization. Pass to penalty in the ProgressiveHedgingSolver factory function. See ?FixedPenalization for parameter descriptions.

StochasticPrograms.ProgressiveHedgingSolvers.AdaptivePenalization — Type

AdaptivePenalization

Functor object for using adaptive penalty in a progressive-hedging algorithm. Create by supplying an Adaptive object through penalty in the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

ζ::T = 0.1: Used to calculate the initial penalty. Non-anticipativity in the initial decision is enforced more as ζ increase.

γ₁::T = 1e-5: Tolerance for primal changes being significant

γ₂::T = 0.01: Tolerance for primal changes dominating dual changes

γ₃::T = 0.25: Tolerance for dual changes dominating primal changes

σ::T = 1e-5: Tolerance for the quadratic penalty dominating the Lagrangian

α::T = 0.95: Penalty decrease after primal changes dominating dual changes

θ::T = 1.1: Penalty increase after dual changes dominating primal changes

ν::T = 0.1: Tolerance for significant non-anticipativity violation

β::T = 1.1: Penalty increase after increased non-anticipativity violation

η::T = 1.25: Default penalty increase in the default case

...

StochasticPrograms.ProgressiveHedgingSolvers.Adaptive — Type

Adaptive

Factory object for AdaptivePenalization. Pass to penalty in the ProgressiveHedgingSolver factory function. See ?AdaptivePenalization for parameter descriptions.