Solver interface

Documentation for StochasticPrograms.jl's interface for structured solvers and sample-based solvers.

Index

StochasticPrograms.AbstractSampledModel

StochasticPrograms.AbstractSampledSolver

StochasticPrograms.AbstractStructuredModel

StochasticPrograms.AbstractStructuredSolver

StochasticPrograms.Asynchronous

StochasticPrograms.InfNorm

StochasticPrograms.Linearized

StochasticPrograms.ManhattanNorm

StochasticPrograms.Quadratic

StochasticPrograms.Serial

StochasticPrograms.Synchronous

StochasticPrograms.SampledModel

StochasticPrograms.StructuredModel

StochasticPrograms.fill_solution!

StochasticPrograms.internal_solver

StochasticPrograms.internal_solver

StochasticPrograms.optimize_sampled!

StochasticPrograms.optimize_structured!

StochasticPrograms.solverstr

StochasticPrograms.solverstr

StochasticPrograms.stochastic_solution

Interface

StochasticPrograms.AbstractStructuredSolver — Type

AbstractStructuredSolver

Abstract supertype for structured solver interface objects.

StochasticPrograms.AbstractStructuredModel — Type

AbstractStructuredModel

Abstract supertype for structured solver objects.

StochasticPrograms.AbstractSampledSolver — Type

AbstractSampledSolver

Abstract supertype for sample-based solver interface objects.

StochasticPrograms.AbstractSampledModel — Type

AbstractSampledModel

Abstract supertype for sample-based solver objects.

StochasticPrograms.SampledModel — Method

SampledModel(stochasticmodel::StochasticModel, sampler::AbstractSampler, solver::AbstractSampledSolver)

Return an instance of AbstractSampledModel based on stochasticmodel, sampler and the given solver.

See also: optimize_sampled!, stochastic_solution

StochasticPrograms.StructuredModel — Method

StructuredModel(stochasticprogram::StochasticProgram, solver::AbstractStructuredSolver)

Return an instance of AbstractStructuredModel based on stochasticprogram and the given solver.

See also: optimize_structured!, fill_solution!

StochasticPrograms.fill_solution! — Method

fill_solution!(stochasticprogram::StochasticProgram, structuredmodel::AbstractStructuredModel)

Fill in the optimal solution in stochasticprogram after a call to optimize_structured!. Should fill in the first stage result and second stage results for each available scenario.

See also: optimize_structured!

StochasticPrograms.internal_solver — Method

internal_solver(solver::AbstractSampledSolver)

Return an AbstractMathProgSolver, if available, from solver.

StochasticPrograms.internal_solver — Method

internal_solver(solver::AbstractStructuredSolver)

Return an AbstractMathProgSolver, if available, from solver.

StochasticPrograms.optimize_sampled! — Method

optimize_sampled!(sampledmodel::AbstractSampledmodel, sampler::AbstractSampler, confidence::AbstractFloat)

Optimize the AbstractSampledModel to the given confidence level, using sampler to generate scenarios. This should approximately optimize the stochasticmodel the sampled model was instantiated from.

See also: stochastic_solution

StochasticPrograms.optimize_structured! — Method

optimize_structured!(structuredmodel::AbstractStructuredModel)

Optimize the AbstractStructuredModel, which also optimizes the stochasticprogram it was instantiated from.

See also: fill_solution!

StochasticPrograms.solverstr — Method

solverstr(solver::AbstractStructuredSolver)

Optionally, return a string identifier of AbstractStructuredSolver.

StochasticPrograms.solverstr — Method

solverstr(solver::AbstractStructuredSolver)

Optionally, return a string identifier of AbstractStructuredSolver.

StochasticPrograms.stochastic_solution — Method

stochastic_solution(sampledmodel::AbstractSampledmodel)

Generate a StochasticSolution from sampledmodel after a call to optimize_sampled!. The solution should include an approximately optimal first-stage decision, an an approximate optimal value and a confidence interval around the true optimum of the original stochastic model.

See also: optimize_sampled!, StochasticSolution

Execution

StochasticPrograms.Serial — Type

Serial

Factory object for LShapedSolvers.SerialExecution/ProgressiveHedgingSolvers.SerialExecution. Passed by default to execution in the LShapedSolver and ProgressiveHedgingSolver factory functions.

StochasticPrograms.Synchronous — Type

Synchronous

Factory object for LShapedSolvers.SynchronousExecution/ProgressiveHedgingSolvers.SynchronousExecution. Pass to execution in the LShapedSolver and ProgressiveHedgingSolver factory functions.

StochasticPrograms.Asynchronous — Type

Asynchronous

Factory object for LShapedSolvers.AsynchronousExecution/ProgressiveHedgingSolvers.AsynchronousExecution. Pass to execution in the LShapedSolver and ProgressiveHedgingSolver factory functions

...

Parameters

max_active::Int = 3: Maximum number of active iterations that run asynchronously.

κ::T = 0.5: Relative amount of finished subproblems required to start a new iterate. Governs the amount of asynchronicity.

...

Penalty terms

StochasticPrograms.Quadratic — Type

Quadratic

Functor object for using a quadratic 2-norm penalty term. Requires an AbstractMathProgSolver capable of solving QP problems. Passed by default through penalty where applicable.

StochasticPrograms.Linearized — Type

Linearized

Functor object for using an approximately quadratic penalty term, through linearization. Pass through penalty where applicable.

...

Parameters

nbreakpoints::Int: Number of cutting planes used to approximate quadratic term

...

StochasticPrograms.InfNorm — Type

InfNorm

Functor object for using a linear ∞-norm penalty term. Pass through penalty where applicable.

StochasticPrograms.ManhattanNorm — Type

ManhattanNorm

Functor object for using a linear 1-norm penalty term. Pass through penalty where applicable.