L-shaped solvers

Documentation for StochasticPrograms.jl's L-shaped solvers.

Index

StochasticPrograms.LShaped.AbstractLShapedAttribute

StochasticPrograms.LShaped.Aggregate

StochasticPrograms.LShaped.AggregationParameter

StochasticPrograms.LShaped.Aggregator

StochasticPrograms.LShaped.AsynchronousExecution

StochasticPrograms.LShaped.ClusterAggregate

StochasticPrograms.LShaped.ClusterAggregation

StochasticPrograms.LShaped.ClusterByReference

StochasticPrograms.LShaped.Consolidate

StochasticPrograms.LShaped.Consolidation

StochasticPrograms.LShaped.ConsolidationParameter

StochasticPrograms.LShaped.Consolidator

StochasticPrograms.LShaped.DontAggregate

StochasticPrograms.LShaped.DontConsolidate

StochasticPrograms.LShaped.DontRegularize

StochasticPrograms.LShaped.DynamicAggregate

StochasticPrograms.LShaped.DynamicAggregation

StochasticPrograms.LShaped.FeasibilityCuts

StochasticPrograms.LShaped.GranulatedAggregate

StochasticPrograms.LShaped.GranulatedAggregation

StochasticPrograms.LShaped.Hierarchical

StochasticPrograms.LShaped.HybridAggregate

StochasticPrograms.LShaped.HybridAggregation

StochasticPrograms.LShaped.Kmedoids

StochasticPrograms.LShaped.LShapedAlgorithm

StochasticPrograms.LShaped.LV

StochasticPrograms.LShaped.LevelSet

StochasticPrograms.LShaped.NoAggregation

StochasticPrograms.LShaped.NoConsolidation

StochasticPrograms.LShaped.NoRegularization

StochasticPrograms.LShaped.Optimizer

StochasticPrograms.LShaped.PartialAggregate

StochasticPrograms.LShaped.PartialAggregation

StochasticPrograms.LShaped.RD

StochasticPrograms.LShaped.RawConsolidationParameter

StochasticPrograms.LShaped.RawRegularizationParameter

StochasticPrograms.LShaped.RegularizationParameter

StochasticPrograms.LShaped.RegularizedDecomposition

StochasticPrograms.LShaped.Regularizer

StochasticPrograms.LShaped.SelectClosest

StochasticPrograms.LShaped.SelectDecaying

StochasticPrograms.LShaped.SelectRandom

StochasticPrograms.LShaped.SelectUniform

StochasticPrograms.LShaped.SerialExecution

StochasticPrograms.LShaped.SortByReference

StochasticPrograms.LShaped.StaticCluster

StochasticPrograms.LShaped.SynchronousExecution

StochasticPrograms.LShaped.TR

StochasticPrograms.LShaped.TrustRegion

StochasticPrograms.LShaped.FullAggregation

StochasticPrograms.LShaped.absolute_distance

StochasticPrograms.LShaped.angular_distance

StochasticPrograms.LShaped.at_tolerance

StochasticPrograms.LShaped.get_aggregation_attribute

StochasticPrograms.LShaped.get_consolidation_attribute

StochasticPrograms.LShaped.get_regularization_attribute

StochasticPrograms.LShaped.set_aggregation_attribute

StochasticPrograms.LShaped.set_aggregation_attribute

StochasticPrograms.LShaped.set_aggregation_attributes

StochasticPrograms.LShaped.set_aggregation_attributes

StochasticPrograms.LShaped.set_consolidation_attribute

StochasticPrograms.LShaped.set_consolidation_attribute

StochasticPrograms.LShaped.set_consolidation_attributes

StochasticPrograms.LShaped.set_consolidation_attributes

StochasticPrograms.LShaped.set_regularization_attribute

StochasticPrograms.LShaped.set_regularization_attribute

StochasticPrograms.LShaped.set_regularization_attributes

StochasticPrograms.LShaped.set_regularization_attributes

StochasticPrograms.LShaped.spatioangular_distance

API

StochasticPrograms.LShaped.LShapedAlgorithm — Type

LShapedAlgorithm

Functor object for the L-shaped algorithm.

...

Algorithm parameters

τ::AbstractFloat = 1e-6: Relative tolerance for convergence checks.

debug::Bool = false: Specifies if extra information should be saved for debugging purposes. Defaults to false for memory efficiency.

cut_scaling::AbstractFloat = 1.0: Rescaling factor for cutting planes to improve numerical stability.

log::Bool = true: Specifices if L-shaped procedure should be logged on standard output or not.

...

StochasticPrograms.LShaped.AbstractLShapedAttribute — Type

AbstractLShapedAttribute

Abstract supertype for attribute objects specific to the L-shaped algorithm.

StochasticPrograms.LShaped.AggregationParameter — Type

AggregationParameter

Abstract supertype for aggregation-specific attributes.

StochasticPrograms.LShaped.Aggregator — Type

Aggregator

An optimizer attribute for specifying an aggregation procedure to be used in the L-shaped algorithm. Options are:

NoAggregation: Multi-cut L-shaped algorithm (default)

PartialAggregation: ?PartialAggregation for parameter descriptions.

FullAggregation: ?FullAggregation for parameter descriptions.

DynamicAggregation: ?DynamicAggregation for parameter descriptions.

ClusterAggregation: ?ClusterAggregation for parameter descriptions.

HybridAggregation: ?HybridAggregation for parameter descriptions.

StochasticPrograms.LShaped.ConsolidationParameter — Type

ConsolidationParameter

Abstract supertype for consolidation-specific attributes.

StochasticPrograms.LShaped.Consolidator — Type

Consolidator

An optimizer attribute for specifying a consolidation procedure to be used in the L-shaped algorithm. Options are:

NoConsolidation

Consolidation

StochasticPrograms.LShaped.FeasibilityCuts — Type

FeasibilityCuts

An optimizer attribute for specifying if feasibility cuts should be used in the L-shaped algorithm. Options are true of false.

StochasticPrograms.LShaped.RegularizationParameter — Type

RegularizationParameter

Abstract supertype for regularization-specific attributes.

StochasticPrograms.LShaped.Regularizer — Type

Regularizer

An optimizer attribute for specifying a regularization procedure to be used in the L-shaped algorithm. Options are:

NoRegularization: L-shaped algorithm (default)

RegularizedDecomposition: Regularized decomposition ?RegularizedDecomposition for parameter descriptions.

TrustRegion: Trust-region ?TrustRegion for parameter descriptions.

LevelSet: Level-set ?LevelSet for parameter descriptions.

StochasticPrograms.LShaped.Optimizer — Type

Optimizer(; <keyword arguments>)

Return an L-shaped optimizer. ...

Arguments

master_optimizer::AbstractOptimizer: MathOptInterface solver capable of solving linear (and possibly quadratic) programs.

subproblem_optimizer::AbstractOptimizer: Optionally specify a different solver for the subproblems.

feasibility_cuts::Bool = false: Specify if feasibility cuts should be used

regularize::AbstractRegularizer = DontRegularize(): Specify regularization procedure (DontRegularize, RegularizedDecomposition/RD/WithRegularizedDecomposition, TrustRegion/TR/WithTrustRegion, LevelSet/LV/WithLevelSets).

aggregate::AbstractAggregator = DontAggregate(): Specify aggregation procedure (DontAggregate, Aggregate, PartialAggregate, DynamicAggregate, ClusterAggregate, GranulatedAggregate, HybridAggregate)

consolidate::AbstractConsolidator = DontConsolidate(): Specify consolidation procedure (DontConsolidate, Consolidate)

execution::Execution = Serial: Specify how algorithm should be executed (Serial, Synchronous, Asynchronous). Distributed variants requires worker cores.

<keyword arguments>: Algorithm specific parameters, See ?LShaped for list of possible arguments and default values.

...

StochasticPrograms.LShaped.get_aggregation_attribute — Method

get_aggregation_attribute(stochasticprogram::StochasticProgram, name::String)

Return the value associated with the aggregation-specific attribute named name in stochasticprogram.

See also: set_aggregation_attribute, set_aggregation_attributes.

StochasticPrograms.LShaped.get_consolidation_attribute — Method

get_consolidation_attribute(stochasticprogram::StochasticProgram, name::String)

Return the value associated with the consolidation-specific attribute named name in stochasticprogram.

See also: set_consolidation_attribute, set_consolidation_attributes.

StochasticPrograms.LShaped.get_regularization_attribute — Method

get_regularization_attribute(stochasticprogram::StochasticProgram, name::String)

Return the value associated with the regularization-specific attribute named name in stochasticprogram.

See also: set_regularization_attribute, set_regularization_attributes.

StochasticPrograms.LShaped.set_aggregation_attribute — Method

set_aggregation_attribute(stochasticprogram::StochasticProgram, name::Union{Symbol, String}, value)

Sets the aggregation-specific attribute identified by name to value.

StochasticPrograms.LShaped.set_aggregation_attributes — Method

set_aggregation_attributes(stochasticprogram::StochasticProgram, pairs::Pair...)

Given a list of attribute => value pairs or a collection of keyword arguments, calls set_aggregation_attribute(stochasticprogram, attribute, value) for each pair.

StochasticPrograms.LShaped.set_consolidation_attribute — Method

set_consolidation_attribute(stochasticprogram::StochasticProgram, name::Union{Symbol, String}, value)

Sets the consolidation-specific attribute identified by name to value.

StochasticPrograms.LShaped.set_consolidation_attributes — Method

set_consolidation_attributes(stochasticprogram::StochasticProgram, pairs::Pair...)

Given a list of attribute => value pairs or a collection of keyword arguments, calls set_consolidation_attribute(stochasticprogram, attribute, value) for each pair.

StochasticPrograms.LShaped.set_regularization_attribute — Method

set_regularization_attribute(stochasticprogram::StochasticProgram, name::Union{Symbol, String}, value)

Sets the regularization-specific attribute identified by name to value.

StochasticPrograms.LShaped.set_regularization_attributes — Method

set_regularization_attributes(stochasticprogram::StochasticProgram, pairs::Pair...)

Given a list of attribute => value pairs or a collection of keyword arguments, calls set_regularization_attribute(stochasticprogram, attribute, value) for each pair.

Execution

StochasticPrograms.LShaped.SerialExecution — Type

SerialExecution

Functor object for using serial execution in a lshaped algorithm. Create by supplying a Serial object through execution in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

StochasticPrograms.LShaped.SynchronousExecution — Type

SynchronousExecution

Functor object for using synchronous execution in an L-shaped algorithm (assuming multiple Julia cores are available). Create by supplying a Synchronous object through execution in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

StochasticPrograms.LShaped.AsynchronousExecution — Type

AsynchronousExecution

Functor object for using synchronous execution in an L-shaped algorithm (assuming multiple Julia cores are available). Create by supplying a Asynchronous object through execution in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

Regularization

StochasticPrograms.LShaped.set_regularization_attribute — Function

set_regularization_attribute(stochasticprogram::StochasticProgram, name::Union{Symbol, String}, value)

Sets the regularization-specific attribute identified by name to value.

StochasticPrograms.LShaped.set_regularization_attributes — Function

set_regularization_attributes(stochasticprogram::StochasticProgram, pairs::Pair...)

Given a list of attribute => value pairs or a collection of keyword arguments, calls set_regularization_attribute(stochasticprogram, attribute, value) for each pair.

StochasticPrograms.LShaped.RawRegularizationParameter — Type

RawRegularizationParameter

An optimizer attribute used for raw parameters of the regularizer. Defers to RawParameter.

StochasticPrograms.LShaped.NoRegularization — Type

NoRegularization

Empty functor object for running an L-shaped algorithm without regularization.

StochasticPrograms.LShaped.DontRegularize — Type

DontRegularize

Factory object for NoRegularization. Passed by default to regularize in the LShapedSolver factory function.

StochasticPrograms.LShaped.RegularizedDecomposition — Type

RegularizedDecomposition

Functor object for using regularized decomposition regularization in an L-shaped algorithm. Create by supplying an RD object through regularize in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

σ::AbstractFloat = 1.0: Initial value of regularization parameter. Controls the relative penalty of the deviation from the current major iterate.

σ̅::AbstractFloat = 4.0: Maximum value of the regularization parameter.

σ̲::AbstractFloat = 0.5: Minimum value of the regularization parameter.

log::Bool = true: Specifices if L-shaped procedure should be logged on standard output or not.

penaltyterm::PenaltyTerm = Quadratic: Specify penaltyterm variant (Quadratic, Linearized, InfNorm, [ManhattanNorm][@ref])

...

StochasticPrograms.LShaped.RD — Type

RD

Factory object for RegularizedDecomposition. Pass to regularize in the LShapedSolver factory function. Equivalent factory calls: RD, WithRD, RegularizedDecomposition, WithRegularizedDecomposition. See ?RegularizedDecomposition for parameter descriptions.

StochasticPrograms.LShaped.TrustRegion — Type

TrustRegion

Functor object for using trust-region regularization in an L-shaped algorithm. Create by supplying a TR object through regularize in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

γ::T = 1e-4: Relative tolerance for deciding if a minor iterate should be accepted as a new major iterate.

Δ::AbstractFloat = 1.0: Initial size of ∞-norm trust-region.

Δ̅::AbstractFloat = 1000.0: Maximum size of ∞-norm trust-region.

...

StochasticPrograms.LShaped.TR — Type

TR

Factory object for TrustRegion. Pass to regularize in the LShapedSolver factory function. Equivalent factory calls: TR, WithTR, TrustRegion, WithTrustRegion. See ?TrustRegion for parameter descriptions.

StochasticPrograms.LShaped.LevelSet — Type

LevelSet

Functor object for using level-set regularization in an L-shaped algorithm. Create by supplying an LV object through regularize in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

λ::AbstractFloat = 0.5: Controls the level position L = (1-λ)θ + λQ̃, a convex combination of the current lower and upper bound.

penaltyterm::PenaltyTerm = Quadratic: Specify penaltyterm variant (Quadratic, Linearized, InfNorm, [ManhattanNorm][@ref])

...

StochasticPrograms.LShaped.LV — Type

LV

Factory object for LevelSet. Pass to regularize in the LShapedSolver factory function. Equivalent factory calls: LV, WithLV, LevelSet, WithLevelSets. See ?LevelSet for parameter descriptions.

Aggregation

StochasticPrograms.LShaped.set_aggregation_attribute — Function

set_aggregation_attribute(stochasticprogram::StochasticProgram, name::Union{Symbol, String}, value)

Sets the aggregation-specific attribute identified by name to value.

StochasticPrograms.LShaped.set_aggregation_attributes — Function

set_aggregation_attributes(stochasticprogram::StochasticProgram, pairs::Pair...)

Given a list of attribute => value pairs or a collection of keyword arguments, calls set_aggregation_attribute(stochasticprogram, attribute, value) for each pair.

StochasticPrograms.LShaped.DontAggregate — Type

DontAggregate

Factory object for NoAggregation. Passed by default to aggregate in LShaped.Optimizer.

StochasticPrograms.LShaped.NoAggregation — Type

NoAggregation

Empty functor object for running an L-shaped algorithm without aggregation (multi-cut L-shaped).

StochasticPrograms.LShaped.Aggregate — Type

Aggregate

Factory object for FullAggregation. Pass to aggregate in LShaped.Optimizer or by setting the Aggregator attribute.

StochasticPrograms.LShaped.PartialAggregate — Type

PartialAggregate

Factory object for PartialAggregation. Pass to aggregate in LShaped.Optimizer or by setting the Aggregator attribute. See ?PartialAggregation for parameter descriptions.

StochasticPrograms.LShaped.PartialAggregation — Type

PartialAggregation

Functor object for using partial aggregation in an L-shaped algorithm. Create by supplying a PartialAggregate object through aggregate in LShaped.Optimizer or by setting the Aggregator attribute.

...

Parameters

size::Int: Number of cuts in each aggregate

...

StochasticPrograms.LShaped.FullAggregation — Method

FullAggregation

Functor object for using complete aggregation in an L-shaped algorithm. Create by supplying an Aggregate object through aggregate in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

StochasticPrograms.LShaped.DynamicAggregate — Type

DynamicAggregate(num_aggregates::Integer, rule::AbstractSelectionRule; lock_after::Function = (τ,n)->false)

Factory object for DynamicAggregation. Pass to aggregate in LShaped.Optimizer or by setting the Aggregator attribute. See ?DynamicAggregation for parameter descriptions.

StochasticPrograms.LShaped.DynamicAggregation — Type

DynamicAggregation

Functor object for using dynamic aggregation in an L-shaped algorithm. Create by supplying a DynamicAggregate object through aggregate in LShaped.Optimizer or by setting the Aggregator attribute.

The following selection rules are available

SelectUniform

SelectDecaying

[SelectRandom](@ref

SelectClosest

SortByReference

...

Parameters

num_aggregates::Int: Number of aggregates

rule::SelectionRule: Rule that determines which aggregate an incoming cut should be placed in

lock_after::Function = (τ,n)->false: Function that determines if the current aggregation scheme should be fixed, based on the current optimality gap τ and the number of iterations n

...

StochasticPrograms.LShaped.ClusterAggregate — Type

ClusterAggregate(rule::AbstractClusterRule; lock_after::Function = (τ,n)->false)

Factory object for ClusterAggregation. Pass to aggregate in LShaped.Optimizer or by setting the Aggregator attribute. See ?ClusterAggregation for parameter descriptions.

StochasticPrograms.LShaped.ClusterAggregation — Type

ClusterAggregation

Functor object for using cluster aggregation in an L-shaped algorithm. Create by supplying a ClusterAggregate object through aggregate in LShaped.Optimizer or by setting the Aggregator attribute.

The following cluster rules are available

StaticCluster

ClusterByReference

[Kmedoids](@ref

Hierarchical

...

Parameters

rule::ClusterRule: Rule that determines how cuts should be sorted into clusters

lock_after::Function = (τ,n)->false: Function that determines if the current aggregation scheme should be fixed, based on the current optimality gap τ and the number of iterations n

...

StochasticPrograms.LShaped.GranulatedAggregate — Type

GranulatedAggregate

Factory object for GranulatedAggregation. Pass to aggregate in LShaped.Optimizer or by setting the Aggregator attribute. See ?GranulatedAggregation for parameter descriptions.

StochasticPrograms.LShaped.GranulatedAggregation — Type

GranulatedAggregation

Functor object for using partial aggregation in an L-shaped algorithm. Create by supplying a GranulatedAggregate object through aggregate in LShaped.Optimizer or by setting the Aggregator attribute.

...

Parameters

size::Int: Number of cuts in each aggregate

...

StochasticPrograms.LShaped.HybridAggregate — Type

HybridAggregate(initial::AbstractAggregator, final::AbstractAggregator, τ::AbstractFloat)

Factory object for HybridAggregation. Pass to aggregate in LShaped.Optimizer or by setting the Aggregator attribute. See ?HybridAggregation for parameter descriptions.

StochasticPrograms.LShaped.HybridAggregation — Type

HybridAggregation

Functor object for using hybrid aggregation in an L-shaped algorithm. Create by supplying a HybridAggregate object through aggregate in LShaped.Optimizer or by setting the Aggregator attribute.

...

Parameters

initial::AbstractAggregator: Initial aggregation scheme

final::AbstractAggregator: Final aggregation scheme

τ::T: The active aggregation scheme is switched from initial to final when the optimality gap decreases below τ

...

Selection rules

StochasticPrograms.LShaped.SelectClosest — Type

SelectClosest(τ::AbstractFloat; distance::Function = absolute_distance)

Incoming cuts are placed into the closest aggregate, according the supplied distance function. An empty aggregate is chosen if no aggregate is within the tolerance τ

The following distance measures are available

absolute_distance

angular_distance

[spatioangular_distance](@ref

StochasticPrograms.LShaped.SelectDecaying — Type

SelectDecaying(T₀::Integer, T̲::Integer = 1, γ::T)

Behaves like SelectUniform, but the uniform aggregate size decays by γ each iteration, starting from T₀. T̲ is an optional lower bound on the aggregate size.

StochasticPrograms.LShaped.SelectRandom — Type

SelectRandom(max = Inf)

Incoming cuts are placed into aggregates randomly. An optional maximum number of cuts max can be specified.

StochasticPrograms.LShaped.SelectUniform — Type

SelectUniform(n::Integer)

Incoming cuts are placed into aggregates uniformly, so that each aggregate has at most n cuts. Behaves as PartialAggregation.

StochasticPrograms.LShaped.SortByReference — Type

SortByReference(τ::AbstractFloat; distance::Function = absolute_distance)

Incoming cuts are placed into an aggregate based on the distance to a reference cut, according the supplied distance function. Behaves as SelectClosest if not withing the tolerance τ to the reference cut.

The following distance measures are available

absolute_distance

angular_distance

[spatioangular_distance](@ref

Cluster rules

StochasticPrograms.LShaped.ClusterByReference — Method

ClusterByReference(τ::AbstractFloat; distance::Function = absolute_distance)

Buffered cuts are aggregated if within the tolerance τ to a reference cut, according the supplied distance function. Behaves as multi-cut otherwise.

The following distance measures are available

absolute_distance

angular_distance

[spatioangular_distance](@ref

StochasticPrograms.LShaped.Hierarchical — Type

Hierarchical(nclusters::Int; distance::Function = absolute_distance, linkage::Symbol = :single)

Buffered cuts are sorted into nclusters clusters, using a Hierarchical algorithm, with the given linkage, over a generalized distance matrix.

The following distance measures are available

absolute_distance

angular_distance

[spatioangular_distance](@ref

StochasticPrograms.LShaped.Kmedoids — Type

Kmedoids(nclusters::Int; distance::Function = absolute_distance)

Buffered cuts are sorted into nclusters clusters, using a K-medoids algorithm over a generalized distance matrix.

The following distance measures are available

absolute_distance

angular_distance

[spatioangular_distance](@ref

StochasticPrograms.LShaped.StaticCluster — Type

StaticCluster(clusters::Vector{Float64})

Buffered cuts are sorting according to the supplied weights clusters

Distance measures

StochasticPrograms.LShaped.absolute_distance — Method

absolute_distance(c₁::AnyOptimalityCut, c₂::AnyOptimalityCut)

Absolute distance between two optimality cuts

StochasticPrograms.LShaped.angular_distance — Method

angular_distance(c₁::AnyOptimalityCut, c₂::AnyOptimalityCut)

Angular distance between two optimality cuts

StochasticPrograms.LShaped.spatioangular_distance — Method

spatioangular_distance(c₁::AnyOptimalityCut, c₂::AnyOptimalityCut)

Spatioangular distance between two optimality cuts.

Consolidation

StochasticPrograms.LShaped.set_consolidation_attribute — Function

set_consolidation_attribute(stochasticprogram::StochasticProgram, name::Union{Symbol, String}, value)

Sets the consolidation-specific attribute identified by name to value.

StochasticPrograms.LShaped.set_consolidation_attributes — Function

set_consolidation_attributes(stochasticprogram::StochasticProgram, pairs::Pair...)

Given a list of attribute => value pairs or a collection of keyword arguments, calls set_consolidation_attribute(stochasticprogram, attribute, value) for each pair.

StochasticPrograms.LShaped.Consolidate — Type

Consolidate

Factory object for Consolidation. Pass to consolidate in the LShapedSolver factory function. See ?Consolidation for parameter descriptions.

StochasticPrograms.LShaped.Consolidation — Type

Consolidation

Functor object for using consolidation in an L-shaped algorithm. Create by supplying a Consolidate object through consolidate in the LShapedSolver factory function and then pass to a StochasticPrograms.jl model.

...

Algorithm parameters

tresh::T = 0.95: Relative amount of redundant cuts in a former iteration required to consider the iteration redundant

at::Int = 5.0: Number of times an iteration can be redundant before consolidation is triggered

rebuild::Function = at_tolerance(): Function deciding when the master model should be rebuilt according to performed consolidations

...

StochasticPrograms.LShaped.DontConsolidate — Type

DontConsolidate

Factory object for NoConsolidation. Passed by default to consolidate in the LShapedSolver factory function.

StochasticPrograms.LShaped.NoConsolidation — Type

NoRegularization

Empty functor object for running the L-shaped algorithm without consolidation.

StochasticPrograms.LShaped.RawConsolidationParameter — Type

RawConsolidationParameter

An optimizer attribute used for raw parameters of the consolidator. Defers to RawParameter.

StochasticPrograms.LShaped.at_tolerance — Method

at_tolerance(τ = 0.4, miniter = 0)

Rebuild master when at least nconsolidations*miniter iterations has passed and the ratio of number of cuts in the consolidated collection and the number of cuts in the master model has decreased below τ.