Progressive-hedging solvers

Documentation for StochasticPrograms.jl's progressive-hedging solvers.

Index

StochasticPrograms.ProgressiveHedging.AbstractProgressiveHedgingAttribute

StochasticPrograms.ProgressiveHedging.Adaptive

StochasticPrograms.ProgressiveHedging.AdaptivePenalization

StochasticPrograms.ProgressiveHedging.AsynchronousExecution

StochasticPrograms.ProgressiveHedging.DualTolerance

StochasticPrograms.ProgressiveHedging.Fixed

StochasticPrograms.ProgressiveHedging.FixedPenalization

StochasticPrograms.ProgressiveHedging.Optimizer

StochasticPrograms.ProgressiveHedging.PenalizationParameter

StochasticPrograms.ProgressiveHedging.Penalizer

StochasticPrograms.ProgressiveHedging.PenaltyTerm

StochasticPrograms.ProgressiveHedging.PrimalTolerance

StochasticPrograms.ProgressiveHedging.ProgressiveHedgingAlgorithm

StochasticPrograms.ProgressiveHedging.RawPenalizationParameter

StochasticPrograms.ProgressiveHedging.SerialExecution

StochasticPrograms.ProgressiveHedging.SynchronousExecution

StochasticPrograms.ProgressiveHedging.get_penalization_attribute

StochasticPrograms.ProgressiveHedging.set_penalization_attribute

StochasticPrograms.ProgressiveHedging.set_penalization_attribute

StochasticPrograms.ProgressiveHedging.set_penalization_attributes

StochasticPrograms.ProgressiveHedging.set_penalization_attributes

API

StochasticPrograms.ProgressiveHedging.ProgressiveHedgingAlgorithm — Type

ProgressiveHedgingAlgorithm

Functor object for the progressive-hedging algorithm. Create using the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

τ::AbstractFloat = 1e-6: Relative tolerance for convergence checks.

log::Bool = true: Specifices if progressive-hedging procedure should be logged on standard output or not.

...

StochasticPrograms.ProgressiveHedging.AbstractProgressiveHedgingAttribute — Type

AbstractProgressiveHedgingAttribute

Abstract supertype for attribute objects specific to the progressive-hedging algorithm.

StochasticPrograms.ProgressiveHedging.DualTolerance — Type

DualTolerance

An optimizer attribute for specifying the dual tolerance in the progressive-hedging algorithm.

StochasticPrograms.ProgressiveHedging.PenalizationParameter — Type

PenalizationParameter

Abstract supertype for penalization-specific attributes.

StochasticPrograms.ProgressiveHedging.Penalizer — Type

Regularizer

An optimizer attribute for specifying a penalization procedure to be used in the progressive-hedging algorithm. Options are:

Fixed: Fixed penalty (default) ?Fixed for parameter descriptions.

Adaptive: Adaptive penalty update ?Adaptive for parameter descriptions.

StochasticPrograms.ProgressiveHedging.PenaltyTerm — Type

PenaltyTerm

An optimizer attribute for specifying what proximal term should be used in subproblemsof the progressive-hedging algorithm. Options are:

Quadratic (default)

Linearized

InfNorm

ManhattanNorm

StochasticPrograms.ProgressiveHedging.PrimalTolerance — Type

PrimalTolerance

An optimizer attribute for specifying the primal tolerance in the progressive-hedging algorithm.

StochasticPrograms.ProgressiveHedging.Optimizer — Type

Optimizer(; <keyword arguments>)

Return a progressive-hedging optimizer.

...

Arguments

subproblem_optimizer::AbstractOptimizer: MathOptInterface solver capable of solving quadratic programs.

penalty::AbstractPenalizer = Fixed(): Specify penalty update procedure (Fixed, Adaptive)

execution::AbstractExecuter = Serial: Specify how algorithm should be executed (Serial, Synchronous, Asynchronous). Distributed variants requires worker cores.

penaltyterm::PenaltyTerm = Quadratic: Specify penaltyterm variant (Quadratic, Linearized, InfNorm, [ManhattanNorm][@ref])

<keyword arguments>: Algorithm specific parameters, consult individual docstrings (see above list) for list of possible arguments and default values.

...

StochasticPrograms.ProgressiveHedging.get_penalization_attribute — Method

get_penalization_attribute(stochasticprogram::StochasticProgram, name::String)

Return the value associated with the penalization-specific attribute named name in stochasticprogram.

See also: set_penalization_attribute, set_penalization_attributes.

StochasticPrograms.ProgressiveHedging.set_penalization_attribute — Method

set_penalization_attribute(stochasticprogram::StochasticProgram, name::Union{Symbol, String}, value)

Sets the penalization-specific attribute identified by name to value.

StochasticPrograms.ProgressiveHedging.set_penalization_attributes — Method

set_penalization_attributes(stochasticprogram::StochasticProgram, pairs::Pair...)

Given a list of attribute => value pairs or a collection of keyword arguments, calls set_penalization_attribute(stochasticprogram, attribute, value) for each pair.

Execution

StochasticPrograms.ProgressiveHedging.SerialExecution — Type

SerialExecution

Functor object for using serial execution in a progressive-hedging algorithm. Create by supplying a Serial object through execution in the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

StochasticPrograms.ProgressiveHedging.SynchronousExecution — Type

SynchronousExecution

Functor object for using synchronous execution in a progressive-hedging algorithm (assuming multiple Julia cores are available). Create by supplying a Synchronous object through execution in the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

StochasticPrograms.ProgressiveHedging.AsynchronousExecution — Type

AsynchronousExecution

Functor object for using asynchronous execution in a progressive-hedging algorithm (assuming multiple Julia cores are available). Create by supplying an Asynchronous object through execution in the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

Penalties

StochasticPrograms.ProgressiveHedging.set_penalization_attribute — Function

set_penalization_attribute(stochasticprogram::StochasticProgram, name::Union{Symbol, String}, value)

Sets the penalization-specific attribute identified by name to value.

StochasticPrograms.ProgressiveHedging.set_penalization_attributes — Function

set_penalization_attributes(stochasticprogram::StochasticProgram, pairs::Pair...)

Given a list of attribute => value pairs or a collection of keyword arguments, calls set_penalization_attribute(stochasticprogram, attribute, value) for each pair.

StochasticPrograms.ProgressiveHedging.RawPenalizationParameter — Type

RawPenalizationParameter

An optimizer attribute used for raw parameters of the penalizer. Defers to RawParameter.

StochasticPrograms.ProgressiveHedging.FixedPenalization — Type

FixedPenalization

Functor object for using fixed penalty in a progressive-hedging algorithm. Create by supplying a Fixed object through penalty in the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

r::T = 1.00: Fixed penalty

...

StochasticPrograms.ProgressiveHedging.Fixed — Type

Fixed

Factory object for FixedPenalization. Pass to penalty in the ProgressiveHedgingSolver factory function. See ?FixedPenalization for parameter descriptions.

StochasticPrograms.ProgressiveHedging.AdaptivePenalization — Type

AdaptivePenalization

Functor object for using adaptive penalty in a progressive-hedging algorithm. Create by supplying an Adaptive object through penalty in the ProgressiveHedgingSolver factory function and then pass to a StochasticPrograms.jl model.

...

Parameters

ζ::T = 0.1: Used to calculate the initial penalty. Non-anticipativity in the initial decision is enforced more as ζ increase.

γ₁::T = 1e-5: Tolerance for primal changes being significant

γ₂::T = 0.01: Tolerance for primal changes dominating dual changes

γ₃::T = 0.25: Tolerance for dual changes dominating primal changes

σ::T = 1e-5: Tolerance for the quadratic penalty dominating the Lagrangian

α::T = 0.95: Penalty decrease after primal changes dominating dual changes

θ::T = 1.1: Penalty increase after dual changes dominating primal changes

ν::T = 0.1: Tolerance for significant non-anticipativity violation

β::T = 1.1: Penalty increase after increased non-anticipativity violation

η::T = 1.25: Default penalty increase in the default case

...

StochasticPrograms.ProgressiveHedging.Adaptive — Type

Adaptive

Factory object for AdaptivePenalization. Pass to penalty in the ProgressiveHedgingSolver factory function. See ?AdaptivePenalization for parameter descriptions.