Public interface

Documentation for StochasticPrograms.jl's public interface.

Contents

Public interface

Contents

Index

Constructors

Scenarios

Model definition

API

Stochastic programming constructs

Index

StochasticPrograms.AbstractSampler

StochasticPrograms.AbstractScenario

StochasticPrograms.ConfidenceInterval

StochasticPrograms.ExpectedScenario

StochasticPrograms.Probability

StochasticPrograms.SAA

StochasticPrograms.SAA

StochasticPrograms.Sampler

StochasticPrograms.Scenario

StochasticPrograms.StochasticProgram

StochasticPrograms.StochasticProgram

StochasticPrograms.StochasticProgram

StochasticPrograms.StochasticProgram

StochasticPrograms.StochasticProgram

StochasticPrograms.StochasticProgram

StochasticPrograms.StochasticSolution

StochasticPrograms.@decision

StochasticPrograms.@expectation

StochasticPrograms.@first_stage

StochasticPrograms.@parameters

StochasticPrograms.@sample

StochasticPrograms.@sampler

StochasticPrograms.@scenario

StochasticPrograms.@second_stage

StochasticPrograms.@stage

StochasticPrograms.@stochastic_model

StochasticPrograms.@uncertain

StochasticPrograms.@zero

StochasticPrograms.DEP

StochasticPrograms.EEV

StochasticPrograms.EEV

StochasticPrograms.EV

StochasticPrograms.EVP

StochasticPrograms.EVPI

StochasticPrograms.EVPI

StochasticPrograms.EVP_decision

StochasticPrograms.EWS

StochasticPrograms.EWS

StochasticPrograms.VRP

StochasticPrograms.VRP

StochasticPrograms.VSS

StochasticPrograms.VSS

StochasticPrograms.WS

StochasticPrograms.WS_decision

StochasticPrograms.add_scenario!

StochasticPrograms.add_scenario!

StochasticPrograms.add_scenarios!

StochasticPrograms.add_scenarios!

StochasticPrograms.add_worker_scenario!

StochasticPrograms.add_worker_scenario!

StochasticPrograms.add_worker_scenarios!

StochasticPrograms.add_worker_scenarios!

StochasticPrograms.confidence_interval

StochasticPrograms.confidence_interval

StochasticPrograms.decision

StochasticPrograms.decision_length

StochasticPrograms.decision_length

StochasticPrograms.deferred

StochasticPrograms.distributed

StochasticPrograms.evaluate_decision

StochasticPrograms.evaluate_decision

StochasticPrograms.evaluate_decision

StochasticPrograms.expected

StochasticPrograms.expected

StochasticPrograms.first_stage_dims

StochasticPrograms.first_stage_nconstraints

StochasticPrograms.gap

StochasticPrograms.generate!

StochasticPrograms.generator

StochasticPrograms.has_generator

StochasticPrograms.instantiate

StochasticPrograms.instantiate

StochasticPrograms.instantiate

StochasticPrograms.lower_bound

StochasticPrograms.masterterms

StochasticPrograms.nscenarios

StochasticPrograms.nstages

StochasticPrograms.nsubproblems

StochasticPrograms.optimal_decision

StochasticPrograms.optimal_decision

StochasticPrograms.optimal_decision

StochasticPrograms.optimal_decision

StochasticPrograms.optimal_value

StochasticPrograms.optimal_value

StochasticPrograms.optimal_value

StochasticPrograms.optimize!

StochasticPrograms.optimize!

StochasticPrograms.outcome_model

StochasticPrograms.outcome_model

StochasticPrograms.probability

StochasticPrograms.probability

StochasticPrograms.probability

StochasticPrograms.recourse_length

StochasticPrograms.sample

StochasticPrograms.sample

StochasticPrograms.sample!

StochasticPrograms.scenario

StochasticPrograms.scenarioproblems

StochasticPrograms.scenarioproblems

StochasticPrograms.scenarios

StochasticPrograms.scenariotext

StochasticPrograms.scenariotype

StochasticPrograms.set_probability!

StochasticPrograms.set_spsolver

StochasticPrograms.spsolver

StochasticPrograms.spsolver_model

StochasticPrograms.stage_model

StochasticPrograms.stage_one_model

StochasticPrograms.stage_parameters

StochasticPrograms.stage_probability

StochasticPrograms.subproblem

StochasticPrograms.subproblems

StochasticPrograms.upper_bound

StochasticPrograms.upper_bound

Constructors

StochasticPrograms.StochasticProgram — Method

StochasticProgram(first_stage_params::Any,
                  second_stage_params::Any,
                  scenarios::Vector{<:AbstractScenario};
                  solver = JuMP.UnsetSolver(),
                  procs = workers())

Create a new two-stage stochastic program with a given collection of scenarios

StochasticPrograms.StochasticProgram — Method

StochasticProgram(scenarios::Vector{<:AbstractScenario};
                  solver = JuMP.UnsetSolver(),
                  procs = workers()) where {SD <: AbstractScenario}

Create a new two-stage stochastic program with a given collection of scenarios and no stage data.

StochasticPrograms.StochasticProgram — Method

StochasticProgram(first_stage_params::Any,
                  second_stage_params::Any,
                  ::Type{S};
                  solver = JuMP.UnsetSolver(),
                  procs = workers()) where S <: AbstractScenario

Create a new two-stage stochastic program with stage data given by first_stage_params and second_stage_params. After construction, scenarios of type S can be added through add_scenario!. Optionally, a capable solver can be supplied to later optimize the stochastic program. If multiple Julia processes are available, the resulting stochastic program will automatically be memory-distributed on these processes. This can be avoided by setting procs = [1].

StochasticPrograms.StochasticProgram — Method

StochasticProgram(first_stage_params::Any,
                  second_stage_params::Any,
                  solver = JuMP.UnsetSolver(),
                  procs = workers())

Create a new two-stage stochastic program with stage data given by first_stage_params and second_stage_params. After construction, scenarios of type Scenario can be added through add_scenario!. Optionally, a capable solver can be supplied to later optimize the stochastic program. If multiple Julia processes are available, the resulting stochastic program will automatically be memory-distributed on these processes. This can be avoided by setting procs = [1].

StochasticPrograms.StochasticProgram — Method

StochasticProgram(::Type{S};
                  solver = JuMP.UnsetSolver(),
                  procs = workers()) where S <: AbstractScenario

Create a new two-stage stochastic program with scenarios of type S and no stage data.

StochasticPrograms.StochasticProgram — Method

StochasticProgram(; solver = JuMP.UnsetSolver(),
                  procs = workers())

Create a new two-stage stochastic program with scenarios of type S and no stage data.

Scenarios

StochasticPrograms.AbstractScenario — Type

AbstractScenario

Abstract supertype for scenario objects.

StochasticPrograms.ExpectedScenario — Type

ExpectedScenario{S <: AbstractScenario}

Wrapper type around an AbstractScenario. Should for convenience be used as the result of a call to expected.

See also expected

StochasticPrograms.Probability — Type

Probability

A type-safe wrapper for Float64 used to represent probability of a scenario occuring.

StochasticPrograms.Scenario — Type

Scenario

Conveniece type that adheres to the AbstractScenario abstraction. Useful when uncertain parameters are a finite set of scalar values.

StochasticPrograms.expected — Method

expected(scenarios::Vector{<:AbstractScenario})

Return the expected scenario out of the collection scenarios in an ExpectedScenario wrapper.

This is defined through classical expectation: sum([probability(s)*s for s in scenarios]), and is always defined for scenarios created through @scenario, if the requested fields support it.

Otherwise, user-defined scenario types must implement this method for full functionality.

See also ExpectedScenario

StochasticPrograms.probability — Method

probability(scenario::AbstractScenario)

Return the probability of scenario occuring.

Is always defined for scenarios created through @scenario. Other user defined scenario types must implement this method to generate a proper probability. The default behaviour is to assume that scenario has a probability field of type Probability

See also: Probability

StochasticPrograms.probability — Method

probability(scenarios::Vector{<:AbstractScenario})

Return the probability of that any scenario in the collection scenarios occurs.

StochasticPrograms.scenariotext — Method

scenariotext(io::IO, scenario::AbstractScenario)

Custom printout called when printing scenario.

StochasticPrograms.set_probability! — Method

set_probability!(scenario::AbstractScenario, probability::AbstractFloat)

Set the probability of scenario occuring.

Is always defined for scenarios created through @scenario. Other user defined scenario types must implement this method.

StochasticPrograms.AbstractSampler — Type

AbstractSampler

Abstract supertype for sampler objects.

StochasticPrograms.Sampler — Type

Sampler

General purpose sampler object that samples Scenario.

See also: Scenario, @sampler

StochasticPrograms.sample — Function

sample(sampler::AbstractSampler{S})

Sample a scenario of type S using sampler.

sample(sampler::AbstractSampler{S}, π::AbstractSampler)

Sample a scenario of type S using sampler and set the probability of the sampled scenario to π.

sample(stochasticmodel::StochasticModel, sampler::AbstractSampler, n::Integer; solver = JuMP.UnsetSolver())

Generate a sampled instance of size n using the model stored in the two-stage stochasticmodel, and the provided sampler.

Optionally, a capable solver can be supplied to sample. Otherwise, any previously set solver will be used.

See also: sample!

StochasticPrograms.@scenario — Macro

@scenario(def)

Define a scenario type compatible with StochasticPrograms using the syntax

@scenario name = begin
    ...structdef...

    [@zero begin
        ...
        return zero(scenario)
    end]

    [@expectation begin
        ...
        return expected(scenarios)
     end]
end

The generated type is referenced through name and a default constructor is always generated. This constructor accepts the keyword probability to set the probability of the scenario occuring. Otherwise, any internal variables and specialized constructors are defined in the @scenario block as they would be in any Julia struct.

If possible, a zero method and an expected method will be generated for the defined type. Otherwise, or if the default implementation is not desired, these can be user provided through @zero and @expectation.

The defined scenario type will be available on all Julia processes.

Examples

The following defines a simple scenario $ξ$ with a single value.

@scenario ExampleScenario = begin
    ξ::Float64
end

ExampleScenario(1.0, probability = 0.5)

# output

ExampleScenario with probability 0.5
  ξ: 1.0

See also: @zero, @expectation, @sampler

StochasticPrograms.@zero — Macro

@zero(def)

Define the additive zero scenario inside a @scenario block using the syntax:

@zero begin
    ...
    return zero_scenario
end

Examples

The following defines a zero scenario for the example scenario defined in @scenario

@zero begin
    return ExampleScenario(0.0)
end

See also @scenario

StochasticPrograms.@expectation — Macro

@expectation(def)

Define how to form the expected scenario inside a @scenario block. The scenario collection is accessed through the reserved keyword scenarios.

@zero begin
    ...
    return zero_scenario
end

Examples

The following defines expectation for the example scenario defined in @scenario

@expectation begin
    return ExampleScenario(sum([probability(s)*s.ξ for s in scenarios]))
end

See also @scenario

StochasticPrograms.@sampler — Macro

@sampler(def)

Define a sampler type for some scenariotype compatible with StochasticPrograms using the syntax

@sampler samplername = begin
    ...internals...

    @sample scenariotype begin
        ...
        return scenario
    end
end

Any internal state required by the sampler, as well as any specialized constructor, are defined in the @sampler block as they would be in any Julia struct. Define the sample operation inside the @sample block and specify the scenariotype that the sampler returns. The defined sampler will be available on all Julia processes.

Examples

The following defines a simple dummy sampler, with some internal weight value, for the scenario defined in @scenario, and samples one scenario.

@sampler ExampleSampler = begin
    w::Float64

    ExampleSampler(w::AbstractFloat) = new(w)

    @sample ExampleScenario begin
        @parameters w
        return ExampleScenario(w*randn(), probability = rand())
    end
end
s = ExampleSampler(2.0)
s()

# output

ExampleScenario with probability 0.29
  ξ: 1.48

See also: @sample, @scenario

StochasticPrograms.@sample — Macro

@sample(def)

Define the sample operaton inside a @sampler block, using the syntax

@sample begin
    ...
    return sampled_scenario
end

The sampler object is referenced through the reserved keyword sampler, from which any internals can be accessed.

Model definition

StochasticPrograms.@stage — Macro

@stage(def)

Add a stage model generation recipe to stochasticprogram using the syntax

@stage stage stochasticprogram::StochasticProgram = begin
    @parameters param1 param2 ...
    @decision var1 var2 ...
    @uncertain ξ
    ... JuMPdef ...
    ...
end [defer]

where JuMP syntax is used inside the block to define the stage model. During definition, the second stage model is referenced through the reserved keyword model.

Optionally, give the keyword defer after the to delay generation of the stage model.

Examples

The following defines the first stage model given by:

\[ minimize 100x₁ + 150x₂ s.t x₁ + x₂ ≤ 120 x₁ ≥ 40 x₂ ≥ 20\]

and the second-stage model given by:

\[ minimize q₁(ξ)y₁ + q₂(ξ)y₂ s.t 6y₁ + 10y₂ ≤ 60x₁ 8y₁ + 5y₂ ≤ 60x₂ 0 ≤ y₁ ≤ d₁(ξ) 0 ≤ y₂ ≤ d₂(ξ)\]

where $q₁(ξ), q₂(ξ), d₁(ξ), d₂(ξ)$ depend on the scenario $ξ$ and $x₁, x₂$ are first stage variables. Two scenarios are added so that two second stage models are generated.

ξ₁ = Scenario(q₁ = -24.0, q₂ = -28.0, d₁ = 500.0, d₂ = 100.0, probability = 0.4)
ξ₂ = Scenario(q₁ = -28.0, q₂ = -32.0, d₁ = 300.0, d₂ = 300.0, probability = 0.6)

sp = StochasticProgram([ξ₁, ξ₂])

@stage 1 sp = begin
    @variable(model, x₁ >= 40)
    @variable(model, x₂ >= 20)
    @objective(model, Min, 100*x₁ + 150*x₂)
    @constraint(model, x₁ + x₂ <= 120)
end

@stage 2 sp = begin
    @decision x₁ x₂
    @uncertain q₁ q₂ d₁ d₂
    @variable(model, 0 <= y₁ <= d₁)
    @variable(model, 0 <= y₂ <= d₂)
    @objective(model, Min, q₁*y₁ + q₂*y₂)
    @constraint(model, 6*y₁ + 10*y₂ <= 60*x₁)
    @constraint(model, 8*y₁ + 5*y₂ <= 80*x₂)
end

# output

Stochastic program with:
 * 2 decision variables
 * 2 recourse variables
 * 2 scenarios of type Scenario
Solver is default solver

See also: @parameters, @decision, @uncertain

StochasticPrograms.@first_stage — Macro

@first_stage(def)

Add a first stage model generation recipe to stochasticprogram using the syntax

@first_stage stochasticprogram::StochasticProgram = begin
    ...
end [defer]

where JuMP syntax is used inside the block to define the first stage model. During definition, the first stage model is referenced through the reserved keyword model.

Optionally, give the keyword defer after the to delay generation of the first stage model.

Examples

The following defines the first stage model given by:

\[ minimize 100x₁ + 150x₂ s.t x₁ + x₂ ≤ 120 x₁ ≥ 40 x₂ ≥ 20\]

@first_stage sp = begin
    @variable(model, x₁ >= 40)
    @variable(model, x₂ >= 20)
    @objective(model, Min, 100*x₁ + 150*x₂)
    @constraint(model, x₁ + x₂ <= 120)
end

See also: @second_stage

StochasticPrograms.@second_stage — Macro

@second_stage(def)

Add a second stage model generation recipe to stochasticprogram using the syntax

@second_stage stochasticprogram::StochasticProgram = begin
    @decision var1 var2 ...
    ...
end [defer]

where JuMP syntax is used inside the block to define the second stage model. During definition, the second stage model is referenced through the reserved keyword model.

Optionally, give the keyword defer after the to delay generation of the first stage model.

Examples

The following defines the second stage model given by:

\[ minimize q₁(ξ)y₁ + q₂(ξ)y₂ s.t 6y₁ + 10y₂ ≤ 60x₁ 8y₁ + 5y₂ ≤ 60x₂ 0 ≤ y₁ ≤ d₁(ξ) 0 ≤ y₂ ≤ d₂(ξ)\]

where $q₁(ξ), q₂(ξ), d₁(ξ), d₂(ξ)$ depend on the scenario $ξ$ and $x₁, x₂$ are first stage variables. Two scenarios are added so that two second stage models are generated.

@second_stage sp = begin
    @decision x₁ x₂
    @uncertain q₁ q₂ d₁ d₂
    @variable(model, 0 <= y₁ <= d₁)
    @variable(model, 0 <= y₂ <= d₂)
    @objective(model, Min, q₁*y₁ + q₂*y₂)
    @constraint(model, 6*y₁ + 10*y₂ <= 60*x₁)
    @constraint(model, 8*y₁ + 5*y₂ <= 80*x₂)
end

See also: @first_stage

StochasticPrograms.@decision — Macro

@decision(def)

In a @stage block, annotate each decision taken in the previous stage using the syntax

@decision var1, var2, ...

Examples

@decision x₁, x₂

See also @parameters, @uncertain, @stage

StochasticPrograms.@parameters — Macro

@parameters(def)

Define the problem parameters in a @stage block

@parameters param1, param2, ...

possibly with default values. Any defined parameter without a default value must be supplied as a keyword argument to instantiate or SAA when creating models.

Examples

@parameters d

@parameters begin
    Crops = [:wheat, :corn, :beets]
    Cost = Dict(:wheat=>150, :corn=>230, :beets=>260)
    Budget = 500
end

See also @decision, @uncertain, @stage

StochasticPrograms.@uncertain — Macro

@uncertain(def)

In a @stage block, annotate each uncertain variable using the syntax

@uncertain var1, var2, ...

This assumes that the [Scenario] type is used. Alternatively, user-defined scenarios can be specified by annotating the type. Also, inside a @stochastic_model block, user-defined scenarios can be created during the @uncertain annotation, following @scenario.

Examples

@uncertain q₁ q₂ d₁ d₂

@scenario Simple = begin
    q₁::Float64
    q₂::Float64
    d₁::Float64
    d₂::Float64
end
@uncertain ξ::SimpleScenario

@stochastic_model begin
    ...
    @uncertain ξ::SimpleScenario = begin
        q₁::Float64
        q₂::Float64
        d₁::Float64
        d₂::Float64
    end
    ...
end

See also @scenario, @parameters, @decision, @stage

StochasticPrograms.@stochastic_model — Macro

@stochastic_model(def)

Define a stochastic model capable of instantiating stochastic programs, using the syntax

sm = @stochastic_model begin
    ...
    @stage x begin
      ...
    end
    ...
end

where the inner blocks are @stage blocks. At least two stages must be specified in consecutive order. A stochastic model object can later be used to instantiate stochastic programs using a given set of scenarios or to create SAA models using samplers.

Examples

The following defines a stochastic model consisitng of the first stage model given by:

\[ minimize 100x₁ + 150x₂ s.t x₁ + x₂ ≤ 120 x₁ ≥ 40 x₂ ≥ 20\]

and the second-stage model given by:

\[ minimize q₁(ξ)y₁ + q₂(ξ)y₂ s.t 6y₁ + 10y₂ ≤ 60x₁ 8y₁ + 5y₂ ≤ 60x₂ 0 ≤ y₁ ≤ d₁(ξ) 0 ≤ y₂ ≤ d₂(ξ)\]

where $q₁(ξ), q₂(ξ), d₁(ξ), d₂(ξ)$ depend on the scenario $ξ$.

sm = @stochastic_model begin
    @stage 1 begin
        @variable(model, x₁ >= 40)
        @variable(model, x₂ >= 20)
        @objective(model, Min, 100*x₁ + 150*x₂)
        @constraint(model, x₁ + x₂ <= 120)
    end
    @stage 2 begin
        @decision x₁ x₂
        @uncertain q₁ q₂ d₁ d₂
        @variable(model, 0 <= y₁ <= d₁)
        @variable(model, 0 <= y₂ <= d₂)
        @objective(model, Min, q₁*y₁ + q₂*y₂)
        @constraint(model, 6*y₁ + 10*y₂ <= 60*x₁)
        @constraint(model, 8*y₁ + 5*y₂ <= 80*x₂)
    end
end

See also: @stage, @parameters, @decision, @uncertain

API

StochasticPrograms.add_scenario! — Function

add_scenario!(scenariogenerator::Function, stochasticprogram::StochasticProgram, stage::Integer = 2; defer::Bool = false)

Store the second stage scenario returned by scenariogenerator in the second stage of the stochasticprogram. Defaults to the second stage.

If defer is true, then model creation is deferred until generate!(stochasticprogram) is called. If the stochasticprogram is distributed, the scenario will be defined on the node that currently has the fewest scenarios.

StochasticPrograms.add_scenario! — Function

add_scenario!(stochasticprogram::StochasticProgram, scenario::AbstractScenario, stage::Integer = 2; defer::Bool = false)

Store the second stage scenario in the stochasticprogram at stage. Defaults to the second stage.

If defer is true, then model creation is deferred until generate!(stochasticprogram) is called. If the stochasticprogram is distributed, the scenario will be defined on the node that currently has the fewest scenarios.

StochasticPrograms.add_scenarios! — Function

add_scenarios!(stochasticprogram::StochasticProgram, scenarios::Vector{<:AbstractScenario}, stage::Integer = 2; defer::Bool = false)

Store the collection of second stage scenarios in the stochasticprogram at stage. Defaults to the second stage.

If defer is true, then model creation is deferred until generate!(stochasticprogram) is called. If the stochasticprogram is distributed, scenarios will be distributed evenly across workers.

StochasticPrograms.add_scenarios! — Function

add_scenarios!(scenariogenerator::Function, stochasticprogram::StochasticProgram, n::Integer, stage::Integer = 2; defer::Bool = false)

Generate n second-stage scenarios using scenariogeneratorand store in the stochasticprogram at stage. Defaults to the second stage.

If defer is true, then model creation is deferred until generate!(stochasticprogram) is called. If the stochasticprogram is distributed, scenarios will be distributed evenly across workers.

StochasticPrograms.decision_length — Method

decision_length(stochasticprogram::TwoStageStochasticProgram)

Return the length of the first-stage decision of the two-stage stochasticprogram.

StochasticPrograms.decision_length — Method

decision_length(stochasticprogram::StochasticProgram, s::Integer)

Return the length of the decision at stage s in the stochasticprogram.

StochasticPrograms.deferred — Method

deferred(stochasticprogram::StochasticProgram)

Return true if stochasticprogram is not fully generated.

StochasticPrograms.distributed — Function

distributed(stochasticprogram::StochasticProgram, s::Integer = 2)

Return true if the stochasticprogram is memory distributed at stage s. Defaults to the second stage.

StochasticPrograms.expected — Function

expected(stochasticprogram::StochasticProgram, s::Integer = 2)

Return the exected scenario of all scenarios of the stochasticprogram at stage s. Defaults to the second stage.

StochasticPrograms.first_stage_dims — Method

first_stage_dims(stochasticprogram::StochasticProgram)

Return a the number of variables and the number of constraints in the the first stage of stochasticprogram as a tuple.

StochasticPrograms.first_stage_nconstraints — Method

first_stage_nconstraints(stochasticprogram::StochasticProgram)

Return the number of constraints in the the first stage of stochasticprogram.

StochasticPrograms.generator — Method

generator(stochasticprogram::StochasticProgram, key::Symbol)

Return the problem generator associated with key in stochasticprogram.

StochasticPrograms.has_generator — Method

has_generator(stochasticprogram::StochasticProgram, key::Symbol)

Return true if a problem generator with key exists in stochasticprogram.

StochasticPrograms.instantiate — Method

instantiate(stochasticmodel::StochasticModel{2},
            scenarios::Vector{<:AbstractScenario};
            solver = JuMP.UnsetSolver(),
            procs = workers(),
            kw...)

Instantiate a new two-stage stochastic program using the model definition stored in the two-stage stochasticmodel, and the given collection of scenarios.

StochasticPrograms.instantiate — Method

instantiate(stochasticmodel::StochasticModel,
            scenarios::Vector{<:AbstractScenario};
            solver = JuMP.UnsetSolver(),
            procs = workers(),
            kw...)

Instantiate a new stochastic program using the model definition stored in stochasticmodel, and the given collection of scenarios.

StochasticPrograms.instantiate — Method

instantiate(stochasticmodel::StochasticModel{2};
            scenariotype::Type{S} = Scenario,
            solver = JuMP.UnsetSolver(),
            procs = workers(),
            kw...) where S <: AbstractScenario

Instantiate a deferred two-stage stochastic program using the model definition stored in the two-stage stochasticmodel over the scenario type S.

StochasticPrograms.masterterms — Function

masterterms(stochasticprogram::StochasticProgram, i::Integer, s::Integer = 2)

Return the first stage terms appearing in scenario i in the stochasticprogram at stage s. Defaults to the second stage.

The master terms are given in sparse format as an array of tuples (row,col,coeff) which specify the occurance of master problem variables in the second stage constraints.

StochasticPrograms.nscenarios — Function

nscenarios(stochasticprogram::StochasticProgram, s::Integer = 2)

Return the number of scenarios in the stochasticprogram at stage s. Defaults to the second stage.

StochasticPrograms.nstages — Method

nstages(stochasticprogram::StochasticProgram)

Return the number of stages in stochasticprogram.

StochasticPrograms.nsubproblems — Function

nsubproblems(stochasticprogram::StochasticProgram, s::Integer = 2)

Return the number of subproblems in the stochasticprogram at stage s. Defaults to the second stage.

StochasticPrograms.optimal_decision — Method

optimal_decision(stochasticprogram::StochasticProgram, var::Symbol)

Return the optimal second stage variable var of stochasticprogram in the ith scenario, after a call to optimize!(stochasticprogram).

StochasticPrograms.optimal_decision — Method

optimal_decision(stochasticprogram::StochasticProgram, i::Integer)

Return the optimal second stage decision of stochasticprogram in the ith scenario, after a call to optimize!(stochasticprogram).

StochasticPrograms.optimal_decision — Method

optimal_decision(stochasticprogram::StochasticProgram, var::Symbol)

Return the optimal first stage variable var of stochasticprogram, after a call to optimize!(stochasticprogram).

StochasticPrograms.optimal_decision — Method

optimal_decision(stochasticprogram::StochasticProgram)

Return the optimal first stage decision of stochasticprogram, after a call to optimize!(stochasticprogram).

StochasticPrograms.optimal_value — Method

optimal_value(stochasticprogram::StochasticProgram, i::Integer)

Return the optimal value of the ith subproblem in stochasticprogram, after a call to optimize!(stochasticprogram).

StochasticPrograms.optimal_value — Method

optimal_value(stochasticprogram::StochasticProgram)

Return the optimal value of stochasticprogram, after a call to optimize!(stochasticprogram).

StochasticPrograms.optimize! — Method

optimize!(stochasticmodel::StochasticModel, sampler::AbstractSampler; solver::SPSolverType = JuMP.UnsetSolver(), confidence::AbstractFloat = 0.95, kwargs...)

Approximately optimize the stochasticmodel using solver when the underlying scenario distribution is inferred by sampler and return a StochasticSolution with the given confidence level.

See also: StochasticSolution

StochasticPrograms.optimize! — Method

optimize!(stochasticprogram::StochasticProgram; solver::SPSolverType = JuMP.UnsetSolver())

Optimize the stochasticprogram in expectation using solver.

Examples

The following solves the stochastic program sp using the L-shaped algorithm.

using LShapedSolvers
using GLPKMathProgInterface

optimize!(sp, solver = LShapedSolver(:ls, GLPKSolverLP()));

# output

L-Shaped Gap  Time: 0:00:01 (4 iterations)
  Objective:       -855.8333333333339
  Gap:             0.0
  Number of cuts:  7
:Optimal

The following solves the stochastic program sp using GLPK on the extended form.

using GLPKMathProgInterface

optimize!(sp, solver = GLPKSolverLP())

:Optimal

See also: VRP

StochasticPrograms.probability — Function

probability(stochasticprogram::StochasticProgram, i::Integer, s::Integer = 2)

Return the probability of scenario ith scenario in the stochasticprogram at stage s occuring. Defaults to the second stage.

StochasticPrograms.recourse_length — Method

recourse_length(stochasticprogram::TwoStageStochasticProgram)

Return the length of the second-stage decision in the two-stage stochasticprogram.

StochasticPrograms.sample! — Function

sample!(stochasticprogram::StochasticProgram, sampler::AbstractSampler, n::Integer, stage::Integer = 2; defer::Bool = false)

Sample n scenarios using sampler and add to the stochasticprogram at stage. Defaults to the second stage.

If defer is true, then model creation is deferred until generate!(stochasticprogram) is called. If the stochasticprogram is distributed, scenarios will be distributed evenly across workers.

StochasticPrograms.sample — Method

sample(stochasticmodel::StochasticModel, sampler::AbstractSampler, n::Integer; solver = JuMP.UnsetSolver())

Generate a sampled instance of size n using the model stored in the two-stage stochasticmodel, and the provided sampler.

Optionally, a capable solver can be supplied to sample. Otherwise, any previously set solver will be used.

See also: sample!

StochasticPrograms.scenario — Function

scenario(stochasticprogram::StochasticProgram, i::Integer, s::Integer = 2)

Return the ith scenario of stochasticprogramat stages`. Defaults to the second stage.

StochasticPrograms.scenarioproblems — Function

scenarioproblems(stochasticprogram::TwoStageStochasticProgram)

Return the scenario problems in stochasticprogram.

StochasticPrograms.scenarioproblems — Method

scenarioproblems(stochasticprogram::StochasticProgram, s::Integer)

Return the scenario problems at stage s in stochasticprogram.

StochasticPrograms.scenarios — Function

scenarios(stochasticprogram::StochasticProgram, s::Integer = 2)

Return an array of all scenarios of the stochasticprogram at stage s. Defaults to the second stage.

StochasticPrograms.scenariotype — Function

scenariotype(stochasticprogram::StochasticProgram, s::Integer = 2)

Return the type of the scenario structure associated with stochasticprogram at stage s. Defaults to the second stage.

StochasticPrograms.set_spsolver — Method

set_spsolver(stochasticprogram::StochasticProgram, spsolver::Union{MathProgBase.AbstractMathProgSolver,AbstractStructuredSolver})

Store the stochastic program solver spsolver of the stochasticprogram.

StochasticPrograms.spsolver — Method

spsolver(stochasticprogram::StochasticProgram)

Return the stochastic program solver spsolver of the stochasticprogram.

StochasticPrograms.spsolver_model — Method

internal_model(stochasticprogram::StochasticProgram)

Return the internal solve model object of the stochasticprogram, after a call to optimize!(stochasticprogram).

StochasticPrograms.stage_parameters — Method

stage_parameters(stochasticprogram::StochasticProgram, s::Integer)

Return the parameters at stage s in stochasticprogram.

StochasticPrograms.stage_probability — Function

stage_probability(stochasticprogram::StochasticProgram, s::Integer = 2)

Return the probability of any scenario in the stochasticprogram at stage s occuring. A well defined model should return 1. Defaults to the second stage.

StochasticPrograms.subproblem — Function

subproblem(stochasticprogram::StochasticProgram, i::Integer, s::Integer = 2)

Return the ith subproblem of the stochasticprogram at stage s. Defaults to the second stage.

StochasticPrograms.subproblems — Function

subproblems(stochasticprogram::StochasticProgram, s::Integer = 2)

Return an array of all subproblems of the stochasticprogram at stage s. Defaults to the second stage.

StochasticPrograms.ConfidenceInterval — Type

ConfidenceInterval

An estimated interval around an unknown quantity at a given confidence level

StochasticPrograms.StochasticSolution — Type

StochasticSolution

Approximate solution to a stochastic model to a given level of confidence.

StochasticPrograms.confidence_interval — Method

confidence_interval(solution::StochasticSolution)

Return a confidence interval around the optimal value of the approximate solution to some stochastic model

StochasticPrograms.decision — Method

decision(solution::StochasticSolution)

Return the optimal first stage decision of the approximate solution to some stochastic model.

StochasticPrograms.optimal_value — Method

optimal_value(solution::StochasticSolution)

Return the optimal value of the approximate solution to some stochastic model

StochasticPrograms.SAA — Type

SAA

Default SampledSolver. Generates a StochasticSolution using the sample average approximation (SAA) method, to the desired confidence level.

StochasticPrograms.SAA — Method

SAA(; solver::SPSolverType = JuMP.UnsetSolver())

Return an SAA where the emerging SAA problems are solved using solver.

StochasticPrograms.generate! — Method

generate!(stochasticprogram::StochasticProgram)

Generate the stochasticprogram using the model definitions from @stage and available data.

StochasticPrograms.outcome_model — Method

outcome_model(stochasticprogram::TwoStageStochasticProgram,
              decision::AbstractVector,
              scenario::AbstractScenario;
              solver::MathProgBase.AbstractMathProgSolver = JuMP.UnsetSolver())

Return the resulting second stage model if decision is the first-stage decision in scenario i, in stochasticprogram. Optionally, supply a capable solver to the outcome model.

StochasticPrograms.outcome_model — Method

outcome_model(stochasticprogram::StochasticProgram{N},
              decisions::NTuple{N-1,AbstractVector}
              scenario_path::NTuple{N-1,AbstractScenario},
              solver::MathProgBase.AbstractMathProgSolver = JuMP.UnsetSolver())

Return the resulting N:th stage model if decisions are the decisions taken in the previous stages and scenario_path are the realized scenarios up to stage N in stochasticprogram. Optionally, supply a capable solver to the outcome model.

StochasticPrograms.stage_one_model — Method

stage_one_model(stochasticprogram::StochasticProgram)

Return a generated copy of the first stage model in stochasticprogram.

StochasticPrograms.confidence_interval — Method

confidence_interval(stochasticmodel::StochasticModel{2},
                    sampler::AbstractSampler;
                    solver = JuMP.UnsetSolver(),
                    confidence = 0.9,
                    N = 100,
                    M = 10,
                    T = 10)

Generate a confidence interval around the true optimum of the two-stage stochasticmodel at level confidence using SAA, over the scenario distribution induced by sampler.

N is the size of the sampled models used to generate the interval and generally governs how tight it is. M is the number of sampled models used in the lower bound calculation, and T is the number of sampled models used in the upper bound calculation.

StochasticPrograms.evaluate_decision — Method

evaluate_decision(stochasticmodel::StochasticModel{2},
                  decision::AbstractVector,
                  sampler::AbstractSampler;
                  solver = JuMP.UnsetSolver(),
                  confidence = 0.95,
                  N = 1000)

Return a statistical estimate of the objective of the two-stage stochasticmodel at decision in the form of a confidence interval at level confidence, over the scenario distribution induced by sampler.

In other words, evaluate decision on a sampled model of size N. Generate an confidence interval using the sample variance of the evaluation.

See also: confidence_interval

StochasticPrograms.evaluate_decision — Method

evaluate_decision(stochasticprogram::TwoStageStochasticProgram,
                  decision::AbstractVector,
                  scenario::AbstractScenario;
                  solver = JuMP.UnsetSolver())

Evaluate the result of taking the first-stage decision if scenario is the actual outcome in stochasticprogram.

StochasticPrograms.evaluate_decision — Method

evaluate_decision(stochasticprogram::TwoStageStochasticProgram,
                  decision::AbstractVector;
                  solver = JuMP.UnsetSolver())

Evaluate the first-stage decision in stochasticprogram.

In other words, evaluate the first-stage objective at decision and solve outcome models of decision for every available scenario. Optionally, supply a capable solver to solve the outcome models. Otherwise, any previously set solver will be used.

StochasticPrograms.gap — Method

gap(stochasticmodel::StochasticModel{2},
    x::AbstractVector,
    sampler::AbstractSampler;
    solver = JuMP.UnsetSolver(),
    confidence = 0.9,
    N = 100,
    M = 10,
    T = 10)

Generate a confidence interval around the gap between the result of using decison x and true optimum of the two-stage stochasticmodel at level confidence using SAA, over the scenario distribution induced by sampler.

N is the size of the SAA models used to generate the interval and generally governs how tight it is. M is the number of sampled models used in the lower bound calculation, and T is the number of sampled models used in the upper bound calculation.

StochasticPrograms.lower_bound — Method

lower_bound(stochasticmodel::StochasticModel{2},
            sampler::AbstractSampler;
            solver = JuMP.UnsetSolver(),
            confidence = 0.95,
            N = 100,
            M = 10)

Generate a confidence interval around a lower bound on the true optimum of the two-stage stochasticmodel at level confidence, over the scenario distribution induced by sampler.

N is the size of the sampled models used to generate the interval and generally governs how tight it is. M is the number of sampled models.

StochasticPrograms.upper_bound — Method

upper_bound(stochasticmodel::StochasticModel{2},
            x::AbstractVector,
            sampler::AbstractSampler;
            solver = JuMP.UnsetSolver(),
            confidence = 0.95,
            T = 10,
            Ñ = 1000)

Generate a confidence interval around an upper bound of the expected value of the decision x in the two-stage stochasticmodel at level confidence, over the scenario distribution induced by sampler.

Ñ is the size of each sampled model and T is the number of sampled models.

StochasticPrograms.upper_bound — Method

upper_bound(stochasticmodel::StochasticModel{2},
            sampler::AbstractSampler;
            solver = JuMP.UnsetSolver(),
            confidence = 0.95,
            N = 100,
            T = 10,
            n = 1000)

Generate a confidence interval around an upper of the true optimum of the two-stage stochasticmodel at level confidence, over the scenario distribution induced by sampler.

N is the size of the sampled model used to generate a candidate decision. Ñ is the size of each sampled model and T is the number of sampled models.

StochasticPrograms.add_worker_scenario! — Function

add_worker_scenario!(scenariogenerator::Function, stochasticprogram::StochasticProgram, w::Integer, stage::Integer = 2; defer::Bool = false)

Store the second stage scenario returned by scenariogenerator in worker node w of the stochasticprogram at stage. Defaults to the second stage.

If defer is true, then model creation is deferred until generate!(stochasticprogram) is called.

StochasticPrograms.add_worker_scenario! — Function

add_worker_scenario!(stochasticprogram::StochasticProgram, scenario::AbstractScenario, w::Integer, stage::Integer = 2; defer::Bool = false)

Store the second stage scenario in worker node w of the stochasticprogram at stage. Defaults to the second stage.

If defer is true, then model creation is deferred until generate!(stochasticprogram) is called.

StochasticPrograms.add_worker_scenarios! — Function

add_worker_scenarios!(stochasticprogram::StochasticProgram, scenarios::Vector{<:AbstractScenario}, w::Integer, stage::Integer = 2; defer::Bool = false)

Store the collection of second stage scenarios in in worker node w of the stochasticprogram at stage. Defaults to the second stage.

If defer is true, then model creation is deferred until generate!(stochasticprogram) is called.

StochasticPrograms.add_worker_scenarios! — Function

add_worker_scenarios!(scenariogenerator::Function, stochasticprogram::StochasticProgram, n::Integer, w::Integer, stage::Integer = 2; defer::Bool = false)

Generate n second-stage scenarios using scenariogeneratorand store them in worker node w of the stochasticprogram at stage. Defaults to the second stage.

If defer is true, then model creation is deferred until generate!(stochasticprogram) is called.

StochasticPrograms.stage_model — Method

stage_two_model(stochasticprogram::StochasticProgram)

Return a generated second stage model corresponding to scenario, in stochasticprogram.

Stochastic programming constructs

StochasticPrograms.DEP — Method

DEP(stochasticprogram::TwoStageStochasticProgram; solver = JuMP.UnsetSolver())

Generate the deterministically equivalent problem (DEP) of the two-stage stochasticprogram.

In other words, generate the extended form the stochasticprogram as a single JuMP model. Optionally, a capable solver can be supplied to DEP. Otherwise, any previously set solver will be used.

See also: VRP, WS

StochasticPrograms.EEV — Method

EEV(stochasticmodel::StochasticModel{2}, sampler::AbstractSampler; solver = JuMP.UnsetSolver(), confidence = 0.95, N::Integer = 100, Ñ::Integer = 1000)

Approximately calculate the expected value of the expected value decision (EEV) of the two-stage stochasticmodel to the given confidence level, over the scenario distribution induced by sampler.

Supply a capable solver to solve the intermediate problems. N is the number of scenarios to sample in order to determine the EVP decision and Ñ is the number of samples in the out-of-sample evaluation of the EVP decision.

See also: EVP, EV

StochasticPrograms.EEV — Method

EEV(stochasticprogram::TwoStageStochasticProgram; solver = JuMP.UnsetSolver())

Calculate the expected value of the expected value solution (EEV) of the two-stage stochasticprogram.

In other words, evaluate the EVP decision. Optionally, supply a capable solver to solve the intermediate problems. The default behaviour is to rely on any previously set solver.

See also: EVP, EV

StochasticPrograms.EV — Method

EV(stochasticprogram::TwoStageStochasticProgram; solver = JuMP.UnsetSolver())

Calculate the optimal value of the EVP of the two-stage stochasticprogram.

Optionally, supply a capable solver to solve the expected value problem. The default behaviour is to rely on any previously set solver.

See also: EVP, EVP_decision, EEV

StochasticPrograms.EVP — Method

EVP(stochasticprogram::TwoStageStochasticProgram; solver = JuMP.UnsetSolver())

Generate the expected value problem (EVP) of the two-stage stochasticprogram.

In other words, generate a wait-and-see model corresponding to the expected scenario over all available scenarios in stochasticprogram. Optionally, supply a capable solver to EVP. Otherwise, any previously set solver will be used.

See also: EVP_decision, EEV, EV, WS

StochasticPrograms.EVPI — Method

EVPI(stochasticmodel::StochasticModel{2}, sampler::AbstractSampler; solver = JuMP.UnsetSolver(), confidence = 0.95)

Approximately calculate the expected value of perfect information (EVPI) of the two-stage stochasticmodel to the given confidence level, over the scenario distribution induced by sampler.

In other words, calculate confidence intervals around VRP and EWS. If they do not overlap, the EVPI is statistically significant, and a confidence interval is calculated and returned. Optionally, supply a capable solver to solve the intermediate problems. Otherwise, any previously set solver will be used.

See also: VRP, EWS, VSS

StochasticPrograms.EVPI — Method

EVPI(stochasticprogram::TwoStageStochasticProgram; solver = JuMP.UnsetSolver())

Calculate the expected value of perfect information (EVPI) of the two-stage stochasticprogram.

In other words, calculate the gap between VRP and EWS. Optionally, supply a capable solver to solve the intermediate problems. Otherwise, any previously set solver will be used.

See also: VRP, EWS, VSS

StochasticPrograms.EVP_decision — Method

EVP_decision(stochasticprogram::TwoStageStochasticProgram; solver = JuMP.UnsetSolver())

Calculate the optimizer of the EVP of the two-stage stochasticprogram.

Optionally, supply a capable solver to solve the expected value problem. The default behaviour is to rely on any previously set solver.

See also: EVP, EV, EEV

StochasticPrograms.EWS — Method

EWS(stochasticmodel::StochasticModel{2}, sampler::AbstractSampler; solver = JuMP.UnsetSolver(), confidence = 0.95, N::Integer = 1000)

Approximately calculate the expected wait-and-see result (EWS) of the two-stage stochasticmodel to the given confidence level, over the scenario distribution induced by sampler.

Supply a capable solver to solve the intermediate problems. N is the number of scenarios to sample.

See also: VRP, WS

StochasticPrograms.EWS — Method

EWS(stochasticprogram::StochasticProgram; solver = JuMP.UnsetSolver())

Calculate the expected wait-and-see result (EWS) of the stochasticprogram.

In other words, calculate the expectated result of all possible wait-and-see models, using the provided scenarios in stochasticprogram. Optionally, a capable solver can be supplied to solve the intermediate problems. Otherwise, any previously set solver will be used.

See also: VRP, WS

StochasticPrograms.VRP — Method

VRP(stochasticmodel::StochasticModel{2}, sampler::AbstractSampler; solver = JuMP.UnsetSolver(), confidence = 0.95)

Return a confidence interval around the value of the recouse problem (VRP) of stochasticmodel to the given confidence level.

Optionally, supply a capable solver to optimize the stochastic program. Otherwise, any previously set solver will be used.

See also: EVPI, VSS, EWS

StochasticPrograms.VRP — Method

VRP(stochasticprogram::StochasticProgram; solver = JuMP.UnsetSolver())

Calculate the value of the recouse problem (VRP) in stochasticprogram.

In other words, optimize the stochastic program and return the optimal value. Optionally, supply a capable solver to optimize the stochastic program. Otherwise, any previously set solver will be used.

See also: EVPI, EWS

StochasticPrograms.VSS — Method

VSS(stochasticmodel::StochasticModel{2}, sampler::AbstractSampler; solver = JuMP.UnsetSolver(), confidence = 0.95, Ñ::Integer = 1000)

Approximately calculate the value of the stochastic solution (VSS) of the two-stage stochasticmodel to the given confidence level, over the scenario distribution induced by sampler.

In other words, calculate confidence intervals around EEV and VRP. If they do not overlap, the VSS is statistically significant, and a confidence interval is calculated and returned. Optionally, supply a capable solver to solve the intermediate problems. Otherwise, any previously set solver will be used. Ñ is the number of samples in the out-of-sample evaluation of EEV.

See also: VRP, EEV, EVPI

StochasticPrograms.VSS — Method

VSS(stochasticprogram::TwoStageStochasticProgram; solver = JuMP.UnsetSolver())

Calculate the value of the stochastic solution (VSS) of the two-stage stochasticprogram.

In other words, calculate the gap between EEV and VRP. Optionally, supply a capable solver to solve the intermediate problems. The default behaviour is to rely on any previously set solver.

StochasticPrograms.WS — Method

WS(stochasticprogram::TwoStageStochasticProgram, scenario::AbstractScenarioaData; solver = JuMP.UnsetSolver())

Generate a wait-and-see (WS) model of the two-stage stochasticprogram, corresponding to scenario.

In other words, generate the first stage and the second stage of the stochasticprogram as if scenario is known to occur. Optionally, a capable solver can be supplied to WS. Otherwise, any previously set solver will be used.

See also: DEP, EVP

StochasticPrograms.WS_decision — Method

WS_decision(stochasticprogram::TwoStageStochasticProgram, scenario::AbstractScenario; solver = JuMP.UnsetSolver())

Calculate the optimizer of the wait-and-see (WS) model of the two-stage stochasticprogram, corresponding to scenario.

Optionally, supply a capable solver to solve the wait-and-see problem. The default behaviour is to rely on any previously set solver.

See also: WS