Solver interface

Solver interface

Documentation for StochasticPrograms.jl's interface for structured solvers.

Index

StochasticPrograms.AbstractStructuredModel

StochasticPrograms.AbstractStructuredSolver

StochasticPrograms.StructuredModel

StochasticPrograms.fill_solution!

StochasticPrograms.internal_solver

StochasticPrograms.optimize_structured!

StochasticPrograms.solverstr

Interface

StochasticPrograms.AbstractStructuredSolver — Type.

AbstractScenario

Abstract supertype for structured solver interface objects.

source

StochasticPrograms.AbstractStructuredModel — Type.

AbstractStructuredModel

Abstract supertype for structured solver objects.

source

StochasticPrograms.StructuredModel — Method.

StructuredModel(stochasticprogram::StochasticProgram, solver::AbstractStructuredSolver)

Return an instance of AbstractStructuredModel based on stochasticprogram and the given solver.

See also: optimize_structured!, fill_solution!

source

StochasticPrograms.fill_solution! — Method.

fill_solution!(stochasticprogram::StochasticProgram, structuredmodel::AbstractStructuredModel)

Fill in the optimal solution in stochasticprogram after a call to optimize_structured!. Should fill in the first stage result and second stage results for each available scenario.

See also: optimize_structured!

source

StochasticPrograms.internal_solver — Method.

internal_solver(solver::AbstractStructuredSolver)

Return an AbstractMathProgSolver, if available, from solver.

source

StochasticPrograms.optimize_structured! — Method.

optimize_structured!(structuredmodel::AbstractStructuredModel)

Optimize the AbstractStructuredModel, which also optimizes the stochasticprogram it was instansiated from.

See also: fill_solution!

source

StochasticPrograms.solverstr — Method.

solverstr(solver::AbstractStructuredModel)

Optionally, return a string identifier of AbstractStructuredModel.

source